Split TerminalHostState implementation
