+++ apps/mac/supaterm/Features/Terminal/TerminalHostState.swift
@@ -8,7 +8,7 @@ import SupatermSupport
 import SupatermTerminalCore
 import SwiftUI
 
-private func normalizedTerminalAgentDetail(_ detail: String?) -> String? {
+func normalizedTerminalAgentDetail(_ detail: String?) -> String? {
   guard let detail = detail?.trimmingCharacters(in: .whitespacesAndNewlines), !detail.isEmpty
   else {
     return nil
@@ -63,7 +63,7 @@ final class TerminalHostState {
     let createdAt: Date
     var subtitle: String
     var title: String
-    fileprivate let origin: NotificationOrigin
+    let origin: NotificationOrigin
 
     init(
       attentionState: SupatermNotificationAttentionState?,
@@ -82,7 +82,7 @@ final class TerminalHostState {
       )
     }
 
-    fileprivate init(
+    init(
       attentionState: SupatermNotificationAttentionState?,
       body: String,
       createdAt: Date,
@@ -105,13 +105,13 @@ final class TerminalHostState {
     case other
   }
 
-  fileprivate enum NotificationOrigin: Equatable, Sendable {
+  enum NotificationOrigin: Equatable, Sendable {
     case structuredAgent(NotificationSemantic)
     case terminalDesktop
     case generic
   }
 
-  private struct RecentStructuredNotification: Equatable, Sendable {
+  struct RecentStructuredNotification: Equatable, Sendable {
     let recordedAt: Date
     let semantic: NotificationSemantic
     let text: String
@@ -180,45 +180,45 @@ final class TerminalHostState {
   }
 
   @ObservationIgnored
-  private let runtime: GhosttyRuntime?
+  let runtime: GhosttyRuntime?
   @ObservationIgnored
-  private let managesTerminalSurfaces: Bool
+  let managesTerminalSurfaces: Bool
   @ObservationIgnored
   @Shared(.supatermSettings)
-  private var supatermSettings = .default
+  var supatermSettings = .default
   @ObservationIgnored
-  private var eventContinuation: AsyncStream<TerminalClient.Event>.Continuation?
+  var eventContinuation: AsyncStream<TerminalClient.Event>.Continuation?
   @ObservationIgnored
   @Shared(.terminalSpaceCatalog)
-  private var spaceCatalog = TerminalSpaceCatalog.default
+  var spaceCatalog = TerminalSpaceCatalog.default
   @ObservationIgnored
-  private var spaceCatalogObservationTask: Task<Void, Never>?
+  var spaceCatalogObservationTask: Task<Void, Never>?
   @ObservationIgnored
-  private var runtimeConfigObserver: NSObjectProtocol?
+  var runtimeConfigObserver: NSObjectProtocol?
   @ObservationIgnored
-  private var lastAppliedSpaceCatalog = TerminalSpaceCatalog.default
+  var lastAppliedSpaceCatalog = TerminalSpaceCatalog.default
   @ObservationIgnored
   var onSessionChange: @MainActor () -> Void = {}
   @ObservationIgnored
   var onCommandFinished: @MainActor (UUID) -> Void = { _ in }
   let spaceManager = TerminalSpaceManager()
 
-  private var pendingEvents: [TerminalClient.Event] = []
-  private var trees: [TerminalTabID: SplitTree<GhosttySurfaceView>] = [:]
-  private var surfaces: [UUID: GhosttySurfaceView] = [:]
-  private var focusedSurfaceIDByTab: [TerminalTabID: UUID] = [:]
-  private var previousFocusedSurfaceIDByTab: [TerminalTabID: UUID] = [:]
-  private var paneNotifications: [UUID: [PaneNotification]] = [:]
-  private var agentActivityByTab: [TerminalTabID: AgentActivity] = [:]
-  private var agentActivitySurfaceIDByTab: [TerminalTabID: UUID] = [:]
-  private var codexHoverMessagesByTab: [TerminalTabID: [String]] = [:]
-  private var previousSelectedTabIDBySpace: [TerminalSpaceID: TerminalTabID] = [:]
-  private var previousSelectedSpaceID: TerminalSpaceID?
-  private var lastEmittedFocusSurfaceID: UUID?
-  private var runtimeConfigGeneration = 0
-  private var suppressesSessionChanges = 0
+  var pendingEvents: [TerminalClient.Event] = []
+  var trees: [TerminalTabID: SplitTree<GhosttySurfaceView>] = [:]
+  var surfaces: [UUID: GhosttySurfaceView] = [:]
+  var focusedSurfaceIDByTab: [TerminalTabID: UUID] = [:]
+  var previousFocusedSurfaceIDByTab: [TerminalTabID: UUID] = [:]
+  var paneNotifications: [UUID: [PaneNotification]] = [:]
+  var agentActivityByTab: [TerminalTabID: AgentActivity] = [:]
+  var agentActivitySurfaceIDByTab: [TerminalTabID: UUID] = [:]
+  var codexHoverMessagesByTab: [TerminalTabID: [String]] = [:]
+  var previousSelectedTabIDBySpace: [TerminalSpaceID: TerminalTabID] = [:]
+  var previousSelectedSpaceID: TerminalSpaceID?
+  var lastEmittedFocusSurfaceID: UUID?
+  var runtimeConfigGeneration = 0
+  var suppressesSessionChanges = 0
   @ObservationIgnored
-  private var recentStructuredNotificationsBySurfaceID: [UUID: RecentStructuredNotification] = [:]
+  var recentStructuredNotificationsBySurfaceID: [UUID: RecentStructuredNotification] = [:]
 
   var windowActivity = WindowActivityState.inactive
 
@@ -249,116 +249,7 @@ final class TerminalHostState {
     }
   }
 
-  func restorationSnapshot() -> TerminalWindowSession {
-    let spaces = spaces.map { space in
-      let tabSnapshots = spaceManager.tabs(in: space.id).compactMap {
-        tab -> (TerminalTabID, TerminalTabSession)? in
-        guard let session = restorationTabSession(for: tab) else { return nil }
-        return (tab.id, session)
-      }
-      let tabs = tabSnapshots.map(\.1)
-      let selectedTabIndex =
-        spaceManager.selectedTabID(in: space.id).flatMap { selectedTabID in
-          tabSnapshots.firstIndex { $0.0 == selectedTabID }
-        }
-        ?? (tabs.isEmpty ? nil : 0)
-      return TerminalWindowSpaceSession(
-        id: space.id,
-        selectedTabIndex: selectedTabIndex,
-        tabs: tabs
-      )
-    }
-
-    let resolvedSelectedSpaceID =
-      self.selectedSpaceID.flatMap { selectedSpaceID in
-        spaces.contains(where: { $0.id == selectedSpaceID }) ? selectedSpaceID : nil
-      }
-      ?? spaces.first?.id
-      ?? spaceCatalog.defaultSelectedSpaceID
-
-    return TerminalWindowSession(
-      selectedSpaceID: resolvedSelectedSpaceID,
-      spaces: spaces
-    )
-  }
-
-  @discardableResult
-  func restore(from session: TerminalWindowSession) -> Bool {
-    guard managesTerminalSurfaces else { return false }
-    let validSpaceIDs = Set(spaces.map(\.id))
-    guard let session = session.pruned(validSpaceIDs: validSpaceIDs) else { return false }
-
-    return withSessionChangesSuppressed {
-      clearSessionState()
-
-      let sessionsBySpaceID = Dictionary(
-        uniqueKeysWithValues: session.spaces.map { ($0.id, $0) }
-      )
-      var restoredTabIDsBySpaceID: [TerminalSpaceID: [TerminalTabID]] = [:]
-
-      for space in spaces {
-        let restoredTabs =
-          sessionsBySpaceID[space.id]?.tabs.enumerated().map { index, session in
-            restoredTabItem(for: session, at: index)
-          } ?? []
-        restoredTabIDsBySpaceID[space.id] = restoredTabs.map(\.id)
-        let selectedTabID =
-          sessionsBySpaceID[space.id]?.selectedTabIndex.flatMap { index in
-            restoredTabs.indices.contains(index) ? restoredTabs[index].id : nil
-          }
-        _ = spaceManager.restoreTabs(
-          restoredTabs,
-          selectedTabID: selectedTabID,
-          in: space.id
-        )
-      }
-
-      for spaceSession in session.spaces {
-        let restoredTabIDs = restoredTabIDsBySpaceID[spaceSession.id] ?? []
-        for (index, tabSession) in spaceSession.tabs.enumerated() {
-          guard restoredTabIDs.indices.contains(index) else { continue }
-          restoreTabSession(
-            tabSession,
-            tabID: restoredTabIDs[index],
-            in: spaceSession.id
-          )
-        }
-      }
-
-      guard spaces.contains(where: { !spaceManager.tabs(in: $0.id).isEmpty }) else {
-        clearSessionState()
-        return false
-      }
-
-      let selectedSpaceID =
-        spaces.contains(where: { $0.id == session.selectedSpaceID })
-        ? session.selectedSpaceID
-        : spaces.first?.id
-      guard let selectedSpaceID, applySelectedSpace(selectedSpaceID) else {
-        clearSessionState()
-        return false
-      }
-
-      if spaceManager.tabs(in: selectedSpaceID).isEmpty {
-        _ = createTab(
-          in: selectedSpaceID,
-          focusing: false,
-          sessionChangesEnabled: false,
-          synchronizesFocus: false
-        )
-      }
-
-      if let selectedTabID {
-        focusSurface(in: selectedTabID)
-      } else {
-        lastEmittedFocusSurfaceID = nil
-      }
-      syncFocus(windowActivity)
-      return true
-    }
-  }
-
-  private func observeRuntimeConfig() {
+  func observeRuntimeConfig() {
     guard let runtime else { return }
     if let runtimeConfigObserver {
       NotificationCenter.default.removeObserver(runtimeConfigObserver)
@@ -427,7 +318,7 @@ final class TerminalHostState {
     }
   }
 
-  private func handleCloseCommand(_ command: TerminalClient.Command) {
+  func handleCloseCommand(_ command: TerminalClient.Command) {
     switch command {
     case .closeSurface(let surfaceID):
       closeSurface(surfaceID)
@@ -448,7 +339,7 @@ final class TerminalHostState {
     }
   }
 
-  private func handleCreationCommand(_ command: TerminalClient.Command) {
+  func handleCreationCommand(_ command: TerminalClient.Command) {
     switch command {
     case .createTab(let inheritingFromSurfaceID):
       _ = createTab(inheritingFromSurfaceID: inheritingFromSurfaceID)
@@ -464,7 +355,7 @@ final class TerminalHostState {
     }
   }
 
-  private func handleInteractionCommand(_ command: TerminalClient.Command) {
+  func handleInteractionCommand(_ command: TerminalClient.Command) {
     switch command {
     case .navigateSearch(let direction):
       _ = navigateSearchOnFocusedSurface(direction)
@@ -485,7 +376,7 @@ final class TerminalHostState {
     }
   }
 
-  private func handleSelectionCommand(_ command: TerminalClient.Command) {
+  func handleSelectionCommand(_ command: TerminalClient.Command) {
     switch command {
     case .selectLastTab:
       selectLastTab()
@@ -518,228 +409,7 @@ final class TerminalHostState {
     }
   }
 
-  var spaces: [TerminalSpaceItem] {
-    spaceManager.spaces
-  }
-
-  var selectedSpaceID: TerminalSpaceID? {
-    spaceManager.selectedSpaceID
-  }
-
-  var tabs: [TerminalTabItem] {
-    spaceManager.tabs
-  }
-
-  var pinnedTabs: [TerminalTabItem] {
-    spaceManager.pinnedTabs
-  }
-
-  var regularTabs: [TerminalTabItem] {
-    spaceManager.regularTabs
-  }
-
-  var visibleTabs: [TerminalTabItem] {
-    spaceManager.visibleTabs
-  }
-
-  var selectedTabID: TerminalTabID? {
-    spaceManager.selectedTabID
-  }
-
-  var selectedTab: TerminalTabItem? {
-    guard let selectedTabID else { return nil }
-    return spaceManager.tab(for: selectedTabID)
-  }
-
-  var selectedTree: SplitTree<GhosttySurfaceView>? {
-    guard let selectedTabID else { return nil }
-    return splitTree(for: selectedTabID)
-  }
-
-  var selectedSurfaceView: GhosttySurfaceView? {
-    guard
-      let selectedTabID,
-      let focusedSurfaceID = focusedSurfaceIDByTab[selectedTabID]
-    else {
-      return nil
-    }
-    return surfaces[focusedSurfaceID]
-  }
-
-  var selectedSurfaceState: GhosttySurfaceState? {
-    selectedSurfaceView?.bridge.state
-  }
-
-  var commandPaletteSnapshot: TerminalCommandPaletteSnapshot {
-    _ = runtimeConfigGeneration
-
-    let ghosttyCommands = runtime?.commandPaletteEntries().filter(\.isSupported) ?? []
-    var ghosttyShortcutDisplayByAction: [String: String] = [:]
-    for command in ghosttyCommands {
-      if let shortcut = runtime?.keyboardShortcut(forAction: command.action)?.display {
-        ghosttyShortcutDisplayByAction[command.action] = shortcut
-      }
-    }
-
-    return .init(
-      ghosttyCommands: ghosttyCommands,
-      ghosttyShortcutDisplayByAction: ghosttyShortcutDisplayByAction,
-      hasFocusedSurface: selectedSurfaceView != nil,
-      selectedSpaceID: selectedSpaceID,
-      spaces: spaces,
-      selectedTabID: selectedTabID,
-      visibleTabs: visibleTabs
-    )
-  }
-
-  func sidebarTerminalProgress(for tabID: TerminalTabID) -> TerminalSidebarTerminalProgress? {
-    Self.sidebarTerminalProgress(
-      state: focusedSurfaceIDByTab[tabID].flatMap { surfaceID in
-        surfaces[surfaceID]?.bridge.state
-      }
-    )
-  }
-
-  var selectedPaneIsZoomed: Bool {
-    Self.isPaneZoomed(
-      focusedSurfaceID: currentFocusedSurfaceID(),
-      in: selectedTree
-    )
-  }
-
-  var selectedPaneDisplayTitle: String {
-    Self.selectedPaneDisplayTitle(
-      focusedSurfaceID: currentFocusedSurfaceID(),
-      in: selectedTree,
-      titleOverride: { $0.bridge.state.titleOverride },
-      title: { $0.bridge.state.title },
-      pwd: { $0.bridge.state.pwd }
-    )
-  }
-
-  func contextSurfaceID(for tabID: TerminalTabID) -> UUID? {
-    if let focusedSurfaceID = focusedSurfaceIDByTab[tabID], surfaces[focusedSurfaceID] != nil {
-      return focusedSurfaceID
-    }
-    return trees[tabID]?.root?.leftmostLeaf().id
-  }
-
-  func paneWorkingDirectories(for tabID: TerminalTabID) -> [String] {
-    Self.paneWorkingDirectories(
-      in: splitTree(for: tabID),
-      pwd: { $0.bridge.state.pwd }
-    )
-  }
-
-  var terminalBackgroundColor: Color {
-    _ = runtimeConfigGeneration
-    return Color(nsColor: runtime?.backgroundColor() ?? .windowBackgroundColor)
-  }
-
-  var notificationAttentionColor: Color {
-    _ = runtimeConfigGeneration
-    return Color(nsColor: runtime?.notificationAttentionColor() ?? .controlAccentColor)
-  }
-
-  func latestNotificationText(for tabID: TerminalTabID) -> String? {
-    Self.notificationText(
-      Self.latestNotification(
-        in: notifications(for: tabID)
-          .values
-          .flatMap { $0 }
-          .filter { $0.attentionState != nil }
-      )
-    )
-  }
-
-  func latestSidebarNotificationPresentation(
-    for tabID: TerminalTabID
-  ) -> SidebarNotificationPresentation? {
-    Self.sidebarNotificationPresentation(
-      Self.latestNotification(
-        in: notifications(for: tabID)
-          .values
-          .flatMap { $0 }
-          .filter { $0.attentionState != nil }
-      )
-    )
-  }
-
-  func notificationRecordCount(for tabID: TerminalTabID) -> Int {
-    notifications(for: tabID)
-      .values
-      .reduce(into: 0) { $0 += $1.count }
-  }
-
-  func unreadNotificationCount(for tabID: TerminalTabID) -> Int {
-    unreadNotifiedSurfaceIDs(in: tabID).count
-  }
-
-  func unreadNotifiedSurfaceIDs(in tabID: TerminalTabID) -> Set<UUID> {
-    Set(
-      notifications(for: tabID)
-        .compactMap { surfaceID, notifications in
-          Self.surfaceAttentionState(in: notifications) == .unread ? surfaceID : nil
-        }
-    )
-  }
-
-  func agentActivity(for tabID: TerminalTabID) -> AgentActivity? {
-    agentActivityByTab[tabID]
-  }
-
-  func codexHoverMarkdown(for tabID: TerminalTabID) -> String? {
-    guard let messages = codexHoverMessagesByTab[tabID], !messages.isEmpty else { return nil }
-    return messages.joined(separator: "\n\n")
-  }
-
-  func showsAgentActivityDetail(for tabID: TerminalTabID) -> Bool {
-    guard let activitySurfaceID = agentActivitySurfaceIDByTab[tabID] else { return false }
-    return focusedSurfaceIDByTab[tabID] == activitySurfaceID
-  }
-
-  @discardableResult
-  func setAgentActivity(_ activity: AgentActivity, for surfaceID: UUID) -> Bool {
-    guard let tabID = tabID(containing: surfaceID) else { return false }
-    agentActivityByTab[tabID] = activity
-    agentActivitySurfaceIDByTab[tabID] = surfaceID
-    return true
-  }
-
-  @discardableResult
-  func clearAgentActivity(for surfaceID: UUID) -> Bool {
-    guard let tabID = tabID(containing: surfaceID) else { return false }
-    agentActivityByTab.removeValue(forKey: tabID)
-    agentActivitySurfaceIDByTab.removeValue(forKey: tabID)
-    return true
-  }
-
-  @discardableResult
-  func clearCodexHoverMessages(for surfaceID: UUID) -> Bool {
-    guard let tabID = tabID(containing: surfaceID) else { return false }
-    codexHoverMessagesByTab.removeValue(forKey: tabID)
-    return true
-  }
-
-  @discardableResult
-  func recordCodexHoverMessages(
-    _ messages: [String],
-    replacing: Bool,
-    for surfaceID: UUID
-  ) -> Bool {
-    guard let tabID = tabID(containing: surfaceID) else { return false }
-    var nextMessages = replacing ? [] : (codexHoverMessagesByTab[tabID] ?? [])
-    for message in messages.compactMap(normalizedTerminalAgentDetail) where nextMessages.last != message {
-      nextMessages.append(message)
-    }
-    if nextMessages.isEmpty {
-      codexHoverMessagesByTab.removeValue(forKey: tabID)
-    } else {
-      codexHoverMessagesByTab[tabID] = nextMessages
-    }
-    return true
-  }
-  private func ensureInitialTab(
+  func ensureInitialTab(
     focusing: Bool,
     startupInput: String? = nil
   ) {
@@ -751,7 +421,7 @@ final class TerminalHostState {
   }
 
   @discardableResult
-  private func createTab(
+  func createTab(
     focusing: Bool = true,
     command: String? = nil,
     initialInput: String? = nil,
@@ -774,7 +444,7 @@ final class TerminalHostState {
   }
 
   @discardableResult
-  private func createTab(
+  func createTab(
     in spaceID: TerminalSpaceID,
     focusing: Bool = true,
     command: String? = nil,
@@ -818,7 +488,7 @@ final class TerminalHostState {
   }
 
   @discardableResult
-  private func applySelectedSpace(_ spaceID: TerminalSpaceID) -> Bool {
+  func applySelectedSpace(_ spaceID: TerminalSpaceID) -> Bool {
     let currentSelectedSpaceID = selectedSpaceID
     guard spaceManager.selectSpace(spaceID) else { return false }
     if currentSelectedSpaceID != spaceID, let currentSelectedSpaceID {
@@ -827,7 +497,7 @@ final class TerminalHostState {
     return true
   }
 
-  private func applySelectedTab(
+  func applySelectedTab(
     _ tabID: TerminalTabID,
     in spaceID: TerminalSpaceID
   ) {
@@ -838,7 +508,7 @@ final class TerminalHostState {
     spaceManager.tabManager(for: spaceID)?.selectTab(tabID)
   }
 
-  private func selectTab(_ tabID: TerminalTabID) {
+  func selectTab(_ tabID: TerminalTabID) {
     guard let space = spaceManager.space(for: tabID) else { return }
     let didChangeSpace = spaceManager.selectedSpaceID != space.id
     guard applySelectedSpace(space.id) else { return }
@@ -851,13 +521,13 @@ final class TerminalHostState {
     sessionDidChange()
   }
 
-  private func selectTab(slot: Int) {
+  func selectTab(slot: Int) {
     let index = slot - 1
     guard visibleTabs.indices.contains(index) else { return }
     selectTab(visibleTabs[index].id)
   }
 
-  private func nextTab() {
+  func nextTab() {
     guard
       let selectedTabID,
       let selectedIndex = tabs.firstIndex(where: { $0.id == selectedTabID }),
@@ -869,11 +539,11 @@ final class TerminalHostState {
     selectTab(tabs[nextIndex].id)
   }
 
-  private func nextSpace() {
+  func nextSpace() {
     selectAdjacentSpace(step: 1)
   }
 
-  private func previousTab() {
+  func previousTab() {
     guard
       let selectedTabID,
       let selectedIndex = tabs.firstIndex(where: { $0.id == selectedTabID }),
@@ -885,27 +555,27 @@ final class TerminalHostState {
     selectTab(tabs[previousIndex].id)
   }
 
-  private func previousSpace() {
+  func previousSpace() {
     selectAdjacentSpace(step: -1)
   }
 
-  private func selectLastTab() {
+  func selectLastTab() {
     guard let selectedSpaceID else { return }
     guard let lastTabID = previousSelectedTabIDBySpace[selectedSpaceID] else { return }
     selectTab(lastTabID)
   }
 
-  private func setPinnedTabOrder(_ orderedIDs: [TerminalTabID]) {
+  func setPinnedTabOrder(_ orderedIDs: [TerminalTabID]) {
     spaceManager.activeTabManager?.setPinnedTabOrder(orderedIDs)
     sessionDidChange()
   }
 
-  private func setRegularTabOrder(_ orderedIDs: [TerminalTabID]) {
+  func setRegularTabOrder(_ orderedIDs: [TerminalTabID]) {
     spaceManager.activeTabManager?.setRegularTabOrder(orderedIDs)
     sessionDidChange()
   }
 
-  private func moveSidebarTab(
+  func moveSidebarTab(
     _ tabID: TerminalTabID,
     pinnedOrder: [TerminalTabID],
     regularOrder: [TerminalTabID]
@@ -916,14 +586,14 @@ final class TerminalHostState {
     sessionDidChange()
   }
 
-  private func togglePinned(_ tabID: TerminalTabID) {
+  func togglePinned(_ tabID: TerminalTabID) {
     spaceManager.space(for: tabID).flatMap { spaceManager.tabManager(for: $0.id) }?.togglePinned(
       tabID)
     sessionDidChange()
   }
 
   @discardableResult
-  private func createSpace(named name: String) throws -> TerminalSpaceID {
+  func createSpace(named name: String) throws -> TerminalSpaceID {
     guard let normalizedName = Self.trimmedNonEmpty(name) else {
       throw TerminalControlError.invalidSpaceName
     }
@@ -947,11 +617,11 @@ final class TerminalHostState {
     return space.id
   }
 
-  private func selectSpace(_ spaceID: TerminalSpaceID) {
+  func selectSpace(_ spaceID: TerminalSpaceID) {
     selectSpace(spaceID, persistDefaultSelection: true)
   }
 
-  private func selectSpace(
+  func selectSpace(
     _ spaceID: TerminalSpaceID,
     persistDefaultSelection: Bool
   ) {
@@ -963,13 +633,13 @@ final class TerminalHostState {
     sessionDidChange()
   }
 
-  private func selectSpace(slot: Int) {
+  func selectSpace(slot: Int) {
     let index = slot == 0 ? 9 : slot - 1
     guard spaces.indices.contains(index) else { return }
     selectSpace(spaces[index].id)
   }
 
-  private func selectAdjacentSpace(step: Int) {
+  func selectAdjacentSpace(step: Int) {
     guard
       spaces.count > 1,
       let selectedSpaceID,
@@ -980,7 +650,7 @@ final class TerminalHostState {
     selectSpace(spaces[targetIndex].id)
   }
 
-  private func renameSpace(_ spaceID: TerminalSpaceID, to name: String) {
+  func renameSpace(_ spaceID: TerminalSpaceID, to name: String) {
     let normalizedName = name.trimmingCharacters(in: .whitespacesAndNewlines)
     guard !normalizedName.isEmpty else { return }
     guard spaceManager.isNameAvailable(normalizedName, excluding: spaceID) else { return }
@@ -991,7 +661,7 @@ final class TerminalHostState {
     _ = writeSpaceCatalog(updatedSpaceCatalog)
   }
 
-  private func deleteSpace(_ spaceID: TerminalSpaceID) {
+  func deleteSpace(_ spaceID: TerminalSpaceID) {
     let remainingSpaces = spaceCatalog.spaces.filter { $0.id != spaceID }
     guard !remainingSpaces.isEmpty else { return }
     guard remainingSpaces.count != spaceCatalog.spaces.count else { return }
@@ -1016,19 +686,19 @@ final class TerminalHostState {
     spaceManager.isNameAvailable(proposedName, excluding: excludedSpaceID)
   }
 
-  private func closeSurface(_ surfaceID: UUID) {
+  func closeSurface(_ surfaceID: UUID) {
     performCloseSurface(surfaceID)
   }
 
-  private func closeTab(_ tabID: TerminalTabID) {
+  func closeTab(_ tabID: TerminalTabID) {
     performCloseTab(tabID)
   }
 
-  private func closeTabs(_ tabIDs: [TerminalTabID]) {
+  func closeTabs(_ tabIDs: [TerminalTabID]) {
     performCloseTabs(tabIDs)
   }
 
-  private func requestCloseSurface(_ surfaceID: UUID, needsConfirmation: Bool? = nil) {
+  func requestCloseSurface(_ surfaceID: UUID, needsConfirmation: Bool? = nil) {
     guard surfaces[surfaceID] != nil else { return }
     emit(
       .closeRequested(
@@ -1040,7 +710,7 @@ final class TerminalHostState {
     )
   }
 
-  private func requestCloseTab(_ tabID: TerminalTabID) {
+  func requestCloseTab(_ tabID: TerminalTabID) {
     guard trees[tabID] != nil else { return }
     emit(
       .closeRequested(
@@ -1052,19 +722,19 @@ final class TerminalHostState {
     )
   }
 
-  private func requestCloseTabsBelow(_ tabID: TerminalTabID) {
+  func requestCloseTabsBelow(_ tabID: TerminalTabID) {
     guard let space = spaceManager.space(for: tabID) else { return }
     guard let tabManager = spaceManager.tabManager(for: space.id) else { return }
     requestCloseTabs(tabManager.tabIDsBelow(tabID))
   }
 
-  private func requestCloseOtherTabs(_ tabID: TerminalTabID) {
+  func requestCloseOtherTabs(_ tabID: TerminalTabID) {
     guard let space = spaceManager.space(for: tabID) else { return }
     guard let tabManager = spaceManager.tabManager(for: space.id) else { return }
     requestCloseTabs(tabManager.otherTabIDs(tabID))
   }
 
-  private func requestCloseTabs(_ tabIDs: [TerminalTabID]) {
+  func requestCloseTabs(_ tabIDs: [TerminalTabID]) {
     let tabIDs = tabIDs.filter { trees[$0] != nil }
     guard !tabIDs.isEmpty else { return }
     emit(
@@ -1078,33 +748,33 @@ final class TerminalHostState {
   }
 
   @discardableResult
-  private func navigateSearchOnFocusedSurface(_ direction: GhosttySearchDirection) -> Bool {
+  func navigateSearchOnFocusedSurface(_ direction: GhosttySearchDirection) -> Bool {
     guard let surface = selectedSurfaceView else { return false }
     surface.navigateSearch(direction)
     return true
   }
 
   @discardableResult
-  private func performBindingActionOnFocusedSurface(_ command: SupatermCommand) -> Bool {
+  func performBindingActionOnFocusedSurface(_ command: SupatermCommand) -> Bool {
     guard let surface = selectedSurfaceView else { return false }
     surface.performBindingAction(command.ghosttyBindingAction)
     return true
   }
 
   @discardableResult
-  private func performGhosttyBindingActionOnFocusedSurface(_ action: String) -> Bool {
+  func performGhosttyBindingActionOnFocusedSurface(_ action: String) -> Bool {
     guard let surface = selectedSurfaceView else { return false }
     surface.performBindingAction(action)
     return true
   }
 
-  private func updateWindowActivity(_ activity: WindowActivityState) {
+  func updateWindowActivity(_ activity: WindowActivityState) {
     windowActivity = activity
     syncFocus(activity)
     clearUnreadOnFocusedSurfaceIfNeeded()
   }
 
-  private func syncFocus(_ activity: WindowActivityState) {
+  func syncFocus(_ activity: WindowActivityState) {
     let selectedTabID = spaceManager.selectedTabID
     var surfaceToFocus: GhosttySurfaceView?
 
@@ -1132,1102 +802,6 @@ final class TerminalHostState {
     }
   }
 
-  func treeSnapshot() -> SupatermTreeSnapshot {
-    let window = SupatermTreeSnapshot.Window(
-      index: 1,
-      isKey: windowActivity.isKeyWindow,
-      spaces: spaces.enumerated().map { spaceOffset, space in
-        let tabs = spaceManager.tabs(in: space.id).enumerated().map { tabOffset, tab in
-          let focusedSurfaceID = focusedSurfaceIDByTab[tab.id]
-          let panes = (trees[tab.id]?.leaves() ?? []).enumerated().map { paneOffset, pane in
-            SupatermTreeSnapshot.Pane(
-              index: paneOffset + 1,
-              id: pane.id,
-              isFocused: pane.id == focusedSurfaceID
-            )
-          }
-
-          return SupatermTreeSnapshot.Tab(
-            index: tabOffset + 1,
-            id: tab.id.rawValue,
-            title: tab.title,
-            isSelected: tab.id == spaceManager.selectedTabID(in: space.id),
-            panes: panes
-          )
-        }
-
-        return SupatermTreeSnapshot.Space(
-          index: spaceOffset + 1,
-          id: space.id.rawValue,
-          name: space.name,
-          isSelected: space.id == selectedSpaceID,
-          tabs: tabs
-        )
-      }
-    )
-    return SupatermTreeSnapshot(windows: [window])
-  }
-
-  func debugWindowSnapshot(index: Int) -> SupatermAppDebugSnapshot.Window {
-    .init(
-      index: index,
-      isKey: windowActivity.isKeyWindow,
-      isVisible: windowActivity.isVisible,
-      spaces: spaces.enumerated().map { spaceOffset, space in
-        let tabs = spaceManager.tabs(in: space.id).enumerated().map { tabOffset, tab in
-          let focusedSurfaceID = focusedSurfaceIDByTab[tab.id]
-          let panes = (trees[tab.id]?.leaves() ?? []).enumerated().map { paneOffset, pane in
-            debugPaneSnapshot(
-              pane,
-              index: paneOffset + 1,
-              isFocused: pane.id == focusedSurfaceID
-            )
-          }
-
-          return SupatermAppDebugSnapshot.Tab(
-            index: tabOffset + 1,
-            id: tab.id.rawValue,
-            title: tab.title,
-            isSelected: tab.id == spaceManager.selectedTabID(in: space.id),
-            isPinned: tab.isPinned,
-            isDirty: tab.isDirty,
-            isTitleLocked: tab.isTitleLocked,
-            hasRunningActivity: panes.contains(where: \.isRunning),
-            hasBell: panes.contains(where: { $0.bellCount > 0 }),
-            hasReadOnly: panes.contains(where: \.isReadOnly),
-            hasSecureInput: panes.contains(where: \.hasSecureInput),
-            panes: panes
-          )
-        }
-
-        return SupatermAppDebugSnapshot.Space(
-          index: spaceOffset + 1,
-          id: space.id.rawValue,
-          name: space.name,
-          isSelected: space.id == selectedSpaceID,
-          tabs: tabs
-        )
-      }
-    )
-  }
-
-  func createPane(_ request: TerminalCreatePaneRequest) throws -> SupatermNewPaneResult {
-    let resolvedTarget = try resolveCreatePaneTarget(request.target)
-    let newSurface = createSurface(
-      tabID: resolvedTarget.tabID,
-      command: request.command,
-      initialInput: nil,
-      inheritingFromSurfaceID: resolvedTarget.anchorSurface.id,
-      workingDirectory: request.cwd.map { URL(fileURLWithPath: $0, isDirectory: true) },
-      context: GHOSTTY_SURFACE_CONTEXT_SPLIT
-    )
-
-    do {
-      let newTree = try resolvedTarget.tree.inserting(
-        view: newSurface,
-        at: resolvedTarget.anchorSurface,
-        direction: mapPaneDirection(request.direction)
-      )
-      let finalTree = request.equalize ? newTree.equalized() : newTree
-      trees[resolvedTarget.tabID] = finalTree
-      updateRunningState(for: resolvedTarget.tabID)
-
-      let nextSelectedTabID = Self.selectedTabID(
-        afterCreatingPaneIn: resolvedTarget.tabID,
-        focusRequested: request.focus,
-        currentSelectedTabID: spaceManager.selectedTabID
-      )
-      if let nextSelectedTabID, nextSelectedTabID != spaceManager.selectedTabID {
-        if let space = spaceManager.space(for: nextSelectedTabID) {
-          _ = applySelectedSpace(space.id)
-          applySelectedTab(nextSelectedTabID, in: space.id)
-        }
-      }
-
-      if request.focus {
-        focusSurface(newSurface, in: resolvedTarget.tabID)
-      }
-
-      syncFocus(windowActivity)
-      sessionDidChange()
-
-      let paneLocation = try resolvedPaneLocation(
-        spaceID: resolvedTarget.spaceID,
-        tabID: resolvedTarget.tabID,
-        surfaceID: newSurface.id,
-        tree: finalTree
-      )
-      let selectionState = Self.newPaneSelectionState(
-        selectedTabID: spaceManager.selectedTabID,
-        targetTabID: resolvedTarget.tabID,
-        windowActivity: windowActivity,
-        focusedSurfaceID: focusedSurfaceIDByTab[resolvedTarget.tabID],
-        surfaceID: newSurface.id
-      )
-
-      return SupatermNewPaneResult(
-        direction: request.direction,
-        isFocused: selectionState.isFocused,
-        isSelectedTab: selectionState.isSelectedTab,
-        windowIndex: 1,
-        spaceIndex: paneLocation.spaceIndex,
-        spaceID: resolvedTarget.spaceID.rawValue,
-        tabIndex: paneLocation.tabIndex,
-        tabID: resolvedTarget.tabID.rawValue,
-        paneIndex: paneLocation.paneIndex,
-        paneID: newSurface.id
-      )
-    } catch let error as TerminalCreatePaneError {
-      newSurface.closeSurface()
-      surfaces.removeValue(forKey: newSurface.id)
-      throw error
-    } catch {
-      newSurface.closeSurface()
-      surfaces.removeValue(forKey: newSurface.id)
-      throw TerminalCreatePaneError.creationFailed
-    }
-  }
-
-  func createTab(_ request: TerminalCreateTabRequest) throws -> SupatermNewTabResult {
-    let resolvedTarget = try resolveCreateTabTarget(request.target)
-    let currentSelectedSpaceID = spaceManager.selectedSpaceID
-    let currentSelectedTabID = spaceManager.selectedTabID
-    var createdTabID: TerminalTabID?
-
-    do {
-      let tabID =
-        createTab(
-          in: resolvedTarget.space.id,
-          focusing: false,
-          command: request.command,
-          initialInput: nil,
-          workingDirectory: request.cwd.map { URL(fileURLWithPath: $0, isDirectory: true) },
-          inheritingFromSurfaceID: resolvedTarget.inheritedSurfaceID,
-          insertion: resolvedNewTabInsertion(anchorTabID: resolvedTarget.anchorTabID),
-          sessionChangesEnabled: false,
-          synchronizesFocus: Self.shouldSyncFocusDuringTabCreation(
-            targetSpaceID: resolvedTarget.space.id,
-            focusRequested: request.focus,
-            currentSelectedSpaceID: currentSelectedSpaceID
-          )
-        )
-      guard
-        let tabID,
-        let tree = trees[tabID],
-        let surfaceID = tree.root?.leftmostLeaf().id
-      else {
-        throw TerminalCreateTabError.creationFailed
-      }
-      createdTabID = tabID
-
-      let resolvedSelectedTabID = Self.selectedTabID(
-        afterCreatingTabIn: resolvedTarget.space.id,
-        targetTabID: tabID,
-        focusRequested: request.focus,
-        currentSelectedSpaceID: currentSelectedSpaceID,
-        currentSelectedTabID: currentSelectedTabID
-      )
-      if let tabManager = spaceManager.tabManager(for: resolvedTarget.space.id),
-        resolvedSelectedTabID != tabManager.selectedTabId
-      {
-        applySelectedTab(resolvedSelectedTabID, in: resolvedTarget.space.id)
-      }
-
-      if request.focus {
-        if currentSelectedSpaceID != resolvedTarget.space.id {
-          selectSpace(resolvedTarget.space.id, persistDefaultSelection: true)
-        }
-        applySelectedTab(tabID, in: resolvedTarget.space.id)
-        if let surface = surfaces[surfaceID] {
-          focusSurface(surface, in: tabID)
-        }
-      }
-
-      syncFocus(windowActivity)
-      sessionDidChange()
-
-      guard
-        let spaceIndex = spaceManager.spaceIndex(for: resolvedTarget.space.id),
-        let tabIndex = spaceManager.tabs(in: resolvedTarget.space.id)
-          .firstIndex(where: { $0.id == tabID }),
-        let paneIndex = tree.leaves().firstIndex(where: { $0.id == surfaceID })
-      else {
-        throw TerminalCreateTabError.creationFailed
-      }
-
-      let selectionState = Self.newTabSelectionState(
-        .init(
-          selectedSpaceID: spaceManager.selectedSpaceID,
-          targetSpaceID: resolvedTarget.space.id,
-          selectedTabID: spaceManager.selectedTabID,
-          targetTabID: tabID,
-          windowActivity: windowActivity,
-          focusedSurfaceID: focusedSurfaceIDByTab[tabID],
-          surfaceID: surfaceID
-        )
-      )
-
-      return .init(
-        isFocused: selectionState.isFocused,
-        isSelectedSpace: selectionState.isSelectedSpace,
-        isSelectedTab: selectionState.isSelectedTab,
-        windowIndex: 1,
-        spaceIndex: spaceIndex,
-        spaceID: resolvedTarget.space.id.rawValue,
-        tabIndex: tabIndex + 1,
-        tabID: tabID.rawValue,
-        paneIndex: paneIndex + 1,
-        paneID: surfaceID
-      )
-    } catch let error as TerminalCreateTabError {
-      if let createdTabID {
-        removeTree(for: createdTabID)
-        spaceManager.tabManager(for: resolvedTarget.space.id)?.closeTab(createdTabID)
-      }
-      throw error
-    } catch {
-      if let createdTabID {
-        removeTree(for: createdTabID)
-        spaceManager.tabManager(for: resolvedTarget.space.id)?.closeTab(createdTabID)
-      }
-      throw TerminalCreateTabError.creationFailed
-    }
-  }
-
-  func notify(_ request: TerminalNotifyRequest) throws -> SupatermNotifyResult {
-    try notify(request, origin: .generic)
-  }
-
-  func notifyStructuredAgent(
-    _ request: TerminalNotifyRequest,
-    semantic: NotificationSemantic
-  ) throws -> SupatermNotifyResult {
-    try notify(request, origin: .structuredAgent(semantic))
-  }
-
-  func focusPane(_ target: TerminalPaneTarget) throws -> SupatermFocusPaneResult {
-    let resolvedTarget = try resolvePaneTarget(target)
-    if selectedSpaceID != resolvedTarget.spaceID {
-      selectSpace(resolvedTarget.spaceID, persistDefaultSelection: true)
-    }
-    applySelectedTab(resolvedTarget.tabID, in: resolvedTarget.spaceID)
-    focusSurface(resolvedTarget.anchorSurface, in: resolvedTarget.tabID)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try focusPaneResult(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: resolvedTarget.tree
-    )
-  }
-
-  func lastPane(_ target: TerminalPaneTarget) throws -> SupatermFocusPaneResult {
-    let resolvedTarget = try resolvePaneTarget(target)
-    guard let lastSurfaceID = previousFocusedSurfaceIDByTab[resolvedTarget.tabID] else {
-      throw TerminalControlError.lastPaneNotFound
-    }
-    guard let lastSurface = surfaces[lastSurfaceID] else {
-      throw TerminalControlError.lastPaneNotFound
-    }
-    if selectedSpaceID != resolvedTarget.spaceID {
-      selectSpace(resolvedTarget.spaceID, persistDefaultSelection: true)
-    }
-    applySelectedTab(resolvedTarget.tabID, in: resolvedTarget.spaceID)
-    focusSurface(lastSurface, in: resolvedTarget.tabID)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try focusPaneResult(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: lastSurfaceID,
-      tree: trees[resolvedTarget.tabID] ?? resolvedTarget.tree
-    )
-  }
-
-  func closePane(_ target: TerminalPaneTarget) throws -> SupatermClosePaneResult {
-    let resolvedTarget = try resolvePaneTarget(target)
-    let result = try paneTarget(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: resolvedTarget.tree
-    )
-    performCloseSurface(resolvedTarget.anchorSurface.id)
-    return result
-  }
-
-  func selectTab(_ target: TerminalTabTarget) throws -> SupatermSelectTabResult {
-    let resolvedTarget = try resolveTabTarget(target)
-    if selectedSpaceID != resolvedTarget.spaceID {
-      selectSpace(resolvedTarget.spaceID, persistDefaultSelection: true)
-    }
-    applySelectedTab(resolvedTarget.tabID, in: resolvedTarget.spaceID)
-    focusSurface(in: resolvedTarget.tabID)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try selectTabResult(for: resolvedTarget.tabID)
-  }
-
-  func closeTab(_ target: TerminalTabTarget) throws -> SupatermCloseTabResult {
-    let resolvedTarget = try resolveTabTarget(target)
-    let result = try tabTarget(for: resolvedTarget.tabID)
-    performCloseTab(resolvedTarget.tabID)
-    return result
-  }
-
-  func sendText(_ request: TerminalSendTextRequest) throws -> SupatermSendTextResult {
-    let resolvedTarget = try resolvePaneTarget(request.target)
-    TerminalControlTrace.write(
-      event: "send_text",
-      fields: [
-        "space_id": resolvedTarget.spaceID.rawValue.uuidString.lowercased(),
-        "tab_id": resolvedTarget.tabID.rawValue.uuidString.lowercased(),
-        "surface_id": resolvedTarget.anchorSurface.id.uuidString.lowercased(),
-        "text_length": String(request.text.count),
-        "text_has_cr": request.text.contains("\r") ? "1" : "0",
-        "text_has_lf": request.text.contains("\n") ? "1" : "0",
-        "text_preview": TerminalControlTrace.preview(request.text),
-      ]
-    )
-    resolvedTarget.anchorSurface.bridge.sendText(request.text)
-    return try paneTarget(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: resolvedTarget.tree
-    )
-  }
-
-  func sendKey(_ request: TerminalSendKeyRequest) throws -> SupatermSendKeyResult {
-    let resolvedTarget = try resolvePaneTarget(request.target)
-    TerminalControlTrace.write(
-      event: "send_key",
-      fields: [
-        "key": request.key.rawValue,
-        "space_id": resolvedTarget.spaceID.rawValue.uuidString.lowercased(),
-        "tab_id": resolvedTarget.tabID.rawValue.uuidString.lowercased(),
-        "surface_id": resolvedTarget.anchorSurface.id.uuidString.lowercased(),
-      ]
-    )
-    resolvedTarget.anchorSurface.bridge.sendKey(request.key)
-    return try paneTarget(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: resolvedTarget.tree
-    )
-  }
-
-  func capturePane(_ request: TerminalCapturePaneRequest) throws -> SupatermCapturePaneResult {
-    let resolvedTarget = try resolvePaneTarget(request.target)
-    guard
-      let text = resolvedTarget.anchorSurface.captureText(
-        scope: request.scope,
-        lines: request.lines
-      )
-    else {
-      throw TerminalControlError.captureFailed
-    }
-    return .init(
-      target: try paneTarget(
-        spaceID: resolvedTarget.spaceID,
-        tabID: resolvedTarget.tabID,
-        surfaceID: resolvedTarget.anchorSurface.id,
-        tree: resolvedTarget.tree
-      ),
-      text: text
-    )
-  }
-
-  func resizePane(_ request: TerminalResizePaneRequest) throws -> SupatermResizePaneResult {
-    let resolvedTarget = try resolvePaneTarget(request.target)
-    guard let node = resolvedTarget.tree.find(id: resolvedTarget.anchorSurface.id) else {
-      throw TerminalControlError.resizeFailed
-    }
-    let newTree = try resolvedTarget.tree.resizing(
-      node: node,
-      by: request.amount,
-      in: mapResizeDirection(request.direction),
-      with: CGRect(origin: .zero, size: resolvedTarget.tree.viewBounds())
-    )
-    trees[resolvedTarget.tabID] = newTree
-    sessionDidChange()
-    return try paneTarget(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: newTree
-    )
-  }
-
-  func setPaneSize(_ request: TerminalSetPaneSizeRequest) throws -> SupatermSetPaneSizeResult {
-    let resolvedTarget = try resolvePaneTarget(request.target)
-    guard let node = resolvedTarget.tree.find(id: resolvedTarget.anchorSurface.id) else {
-      throw TerminalControlError.resizeFailed
-    }
-    let newTree = try resolvedTarget.tree.sizing(
-      node: node,
-      to: request.amount,
-      along: mapPaneAxis(request.axis),
-      unit: mapPaneSizeUnit(request.unit),
-      with: CGRect(origin: .zero, size: resolvedTarget.tree.viewBounds())
-    )
-    trees[resolvedTarget.tabID] = newTree
-    sessionDidChange()
-    return try paneTarget(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: newTree
-    )
-  }
-
-  func renameTab(_ request: TerminalRenameTabRequest) throws -> SupatermRenameTabResult {
-    let resolvedTarget = try resolveTabTarget(request.target)
-    let title = Self.trimmedNonEmpty(request.title)
-    setLockedTabTitle(title, for: resolvedTarget.tabID)
-    return .init(
-      isTitleLocked: title != nil,
-      target: try tabTarget(for: resolvedTarget.tabID)
-    )
-  }
-
-  func equalizePanes(_ request: TerminalEqualizePanesRequest) throws -> SupatermEqualizePanesResult {
-    let resolvedTarget = try resolveTabTarget(request.target)
-    trees[resolvedTarget.tabID] = resolvedTarget.tree.equalized()
-    sessionDidChange()
-    return try tabTarget(for: resolvedTarget.tabID)
-  }
-
-  func tilePanes(_ request: TerminalTilePanesRequest) throws -> SupatermTilePanesResult {
-    let resolvedTarget = try resolveTabTarget(request.target)
-    trees[resolvedTarget.tabID] = resolvedTarget.tree.tiled()
-    sessionDidChange()
-    return try tabTarget(for: resolvedTarget.tabID)
-  }
-
-  func mainVerticalPanes(
-    _ request: TerminalMainVerticalPanesRequest
-  ) throws -> SupatermMainVerticalPanesResult {
-    let resolvedTarget = try resolveTabTarget(request.target)
-    trees[resolvedTarget.tabID] = resolvedTarget.tree.mainVertical()
-    sessionDidChange()
-    return try tabTarget(for: resolvedTarget.tabID)
-  }
-
-  func createSpace(_ request: TerminalCreateSpaceRequest) throws -> SupatermCreateSpaceResult {
-    if let windowIndex = request.target.windowIndex, windowIndex != 1 {
-      throw TerminalControlError.windowNotFound(windowIndex)
-    }
-    let spaceID = try createSpace(named: request.name)
-    return try selectSpaceResult(for: spaceID)
-  }
-
-  func selectSpace(_ target: TerminalSpaceTarget) throws -> SupatermSelectSpaceResult {
-    let resolvedTarget = try resolveSpaceTarget(target)
-    selectSpace(resolvedTarget.space.id, persistDefaultSelection: true)
-    return try selectSpaceResult(for: resolvedTarget.space.id)
-  }
-
-  func closeSpace(_ target: TerminalSpaceTarget) throws -> SupatermCloseSpaceResult {
-    let resolvedTarget = try resolveSpaceTarget(target)
-    let result = try spaceTarget(for: resolvedTarget.space.id)
-    if spaces.count == 1 {
-      throw TerminalControlError.onlyRemainingSpace
-    }
-    deleteSpace(resolvedTarget.space.id)
-    return result
-  }
-
-  func renameSpace(_ request: TerminalRenameSpaceRequest) throws -> SupatermSpaceTarget {
-    let resolvedTarget = try resolveSpaceTarget(request.target)
-    renameSpace(resolvedTarget.space.id, to: request.name)
-    return try spaceTarget(for: resolvedTarget.space.id)
-  }
-
-  func nextSpace(_ request: TerminalSpaceNavigationRequest) throws -> SupatermSelectSpaceResult {
-    let currentSpaceID = try resolvedNavigationSpaceID(request)
-    let allSpaces = spaces
-    guard
-      let currentIndex = allSpaces.firstIndex(where: { $0.id == currentSpaceID }),
-      !allSpaces.isEmpty
-    else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    let nextIndex = (currentIndex + 1) % allSpaces.count
-    selectSpace(allSpaces[nextIndex].id, persistDefaultSelection: true)
-    return try selectSpaceResult(for: allSpaces[nextIndex].id)
-  }
-
-  func previousSpace(_ request: TerminalSpaceNavigationRequest) throws -> SupatermSelectSpaceResult {
-    let currentSpaceID = try resolvedNavigationSpaceID(request)
-    let allSpaces = spaces
-    guard
-      let currentIndex = allSpaces.firstIndex(where: { $0.id == currentSpaceID }),
-      !allSpaces.isEmpty
-    else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    let previousIndex = (currentIndex - 1 + allSpaces.count) % allSpaces.count
-    selectSpace(allSpaces[previousIndex].id, persistDefaultSelection: true)
-    return try selectSpaceResult(for: allSpaces[previousIndex].id)
-  }
-
-  func lastSpace(_ request: TerminalSpaceNavigationRequest) throws -> SupatermSelectSpaceResult {
-    if let windowIndex = request.windowIndex, windowIndex != 1 {
-      throw TerminalControlError.windowNotFound(windowIndex)
-    }
-    guard let previousSelectedSpaceID else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    selectSpace(previousSelectedSpaceID, persistDefaultSelection: true)
-    return try selectSpaceResult(for: previousSelectedSpaceID)
-  }
-
-  func nextTab(_ request: TerminalTabNavigationRequest) throws -> SupatermSelectTabResult {
-    let spaceID = try resolvedNavigationSpaceID(request)
-    let tabs = spaceManager.tabs(in: spaceID)
-    guard !tabs.isEmpty else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    let currentTabID = spaceManager.selectedTabID(in: spaceID) ?? tabs[0].id
-    guard let currentIndex = tabs.firstIndex(where: { $0.id == currentTabID }) else {
-      throw TerminalControlError.lastTabNotFound
-    }
-    let nextIndex = (currentIndex + 1) % tabs.count
-    selectSpace(spaceID, persistDefaultSelection: true)
-    applySelectedTab(tabs[nextIndex].id, in: spaceID)
-    focusSurface(in: tabs[nextIndex].id)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try selectTabResult(for: tabs[nextIndex].id)
-  }
-
-  func previousTab(_ request: TerminalTabNavigationRequest) throws -> SupatermSelectTabResult {
-    let spaceID = try resolvedNavigationSpaceID(request)
-    let tabs = spaceManager.tabs(in: spaceID)
-    guard !tabs.isEmpty else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    let currentTabID = spaceManager.selectedTabID(in: spaceID) ?? tabs[0].id
-    guard let currentIndex = tabs.firstIndex(where: { $0.id == currentTabID }) else {
-      throw TerminalControlError.lastTabNotFound
-    }
-    let previousIndex = (currentIndex - 1 + tabs.count) % tabs.count
-    selectSpace(spaceID, persistDefaultSelection: true)
-    applySelectedTab(tabs[previousIndex].id, in: spaceID)
-    focusSurface(in: tabs[previousIndex].id)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try selectTabResult(for: tabs[previousIndex].id)
-  }
-
-  func lastTab(_ request: TerminalTabNavigationRequest) throws -> SupatermSelectTabResult {
-    let spaceID = try resolvedNavigationSpaceID(request)
-    guard let tabID = previousSelectedTabIDBySpace[spaceID] else {
-      throw TerminalControlError.lastTabNotFound
-    }
-    selectSpace(spaceID, persistDefaultSelection: true)
-    applySelectedTab(tabID, in: spaceID)
-    focusSurface(in: tabID)
-    syncFocus(windowActivity)
-    sessionDidChange()
-    return try selectTabResult(for: tabID)
-  }
-
-  @discardableResult
-  func clearRecentStructuredNotification(for surfaceID: UUID) -> Bool {
-    recentStructuredNotificationsBySurfaceID.removeValue(forKey: surfaceID) != nil
-  }
-
-  private func resolveSpaceTarget(_ target: TerminalSpaceTarget) throws -> ResolvedCreateTabTarget {
-    switch target {
-    case .contextPane(let contextPaneID):
-      do {
-        return try resolveCreateTabTarget(.contextPane(contextPaneID))
-      } catch TerminalCreateTabError.contextPaneNotFound {
-        throw TerminalControlError.contextPaneNotFound
-      } catch TerminalCreateTabError.spaceNotFound(let windowIndex, let spaceIndex) {
-        throw TerminalControlError.spaceNotFound(windowIndex: windowIndex, spaceIndex: spaceIndex)
-      } catch TerminalCreateTabError.windowNotFound(let windowIndex) {
-        throw TerminalControlError.windowNotFound(windowIndex)
-      } catch {
-        throw TerminalControlError.contextPaneNotFound
-      }
-
-    case .space(let windowIndex, let spaceIndex):
-      do {
-        return try resolveCreateTabTarget(.space(windowIndex: windowIndex, spaceIndex: spaceIndex))
-      } catch TerminalCreateTabError.spaceNotFound(let resolvedWindowIndex, let resolvedSpaceIndex) {
-        throw TerminalControlError.spaceNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex
-        )
-      } catch TerminalCreateTabError.windowNotFound(let resolvedWindowIndex) {
-        throw TerminalControlError.windowNotFound(resolvedWindowIndex)
-      } catch {
-        throw TerminalControlError.spaceNotFound(windowIndex: windowIndex, spaceIndex: spaceIndex)
-      }
-    }
-  }
-
-  private func resolveTabTarget(_ target: TerminalTabTarget) throws -> ResolvedCreatePaneTarget {
-    switch target {
-    case .contextPane(let contextPaneID):
-      do {
-        return try resolveCreatePaneTarget(.contextPane(contextPaneID))
-      } catch TerminalCreatePaneError.contextPaneNotFound {
-        throw TerminalControlError.contextPaneNotFound
-      } catch TerminalCreatePaneError.spaceNotFound(let windowIndex, let spaceIndex) {
-        throw TerminalControlError.spaceNotFound(windowIndex: windowIndex, spaceIndex: spaceIndex)
-      } catch TerminalCreatePaneError.tabNotFound(let windowIndex, let spaceIndex, let tabIndex) {
-        throw TerminalControlError.tabNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex
-        )
-      } catch TerminalCreatePaneError.windowNotFound(let windowIndex) {
-        throw TerminalControlError.windowNotFound(windowIndex)
-      } catch {
-        throw TerminalControlError.contextPaneNotFound
-      }
-
-    case .tab(let windowIndex, let spaceIndex, let tabIndex):
-      do {
-        return try resolveCreatePaneTarget(
-          .tab(windowIndex: windowIndex, spaceIndex: spaceIndex, tabIndex: tabIndex)
-        )
-      } catch TerminalCreatePaneError.spaceNotFound(let resolvedWindowIndex, let resolvedSpaceIndex) {
-        throw TerminalControlError.spaceNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex
-        )
-      } catch TerminalCreatePaneError.tabNotFound(
-        let resolvedWindowIndex,
-        let resolvedSpaceIndex,
-        let resolvedTabIndex
-      ) {
-        throw TerminalControlError.tabNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex,
-          tabIndex: resolvedTabIndex
-        )
-      } catch TerminalCreatePaneError.windowNotFound(let resolvedWindowIndex) {
-        throw TerminalControlError.windowNotFound(resolvedWindowIndex)
-      } catch {
-        throw TerminalControlError.tabNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex
-        )
-      }
-    }
-  }
-
-  private func resolvePaneTarget(_ target: TerminalPaneTarget) throws -> ResolvedCreatePaneTarget {
-    switch target {
-    case .contextPane(let contextPaneID):
-      do {
-        return try resolveCreatePaneTarget(.contextPane(contextPaneID))
-      } catch TerminalCreatePaneError.contextPaneNotFound {
-        throw TerminalControlError.contextPaneNotFound
-      } catch TerminalCreatePaneError.spaceNotFound(let windowIndex, let spaceIndex) {
-        throw TerminalControlError.spaceNotFound(windowIndex: windowIndex, spaceIndex: spaceIndex)
-      } catch TerminalCreatePaneError.tabNotFound(let windowIndex, let spaceIndex, let tabIndex) {
-        throw TerminalControlError.tabNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex
-        )
-      } catch TerminalCreatePaneError.paneNotFound(
-        let windowIndex, let spaceIndex, let tabIndex, let paneIndex
-      ) {
-        throw TerminalControlError.paneNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex,
-          paneIndex: paneIndex
-        )
-      } catch TerminalCreatePaneError.windowNotFound(let windowIndex) {
-        throw TerminalControlError.windowNotFound(windowIndex)
-      } catch {
-        throw TerminalControlError.contextPaneNotFound
-      }
-
-    case .pane(let windowIndex, let spaceIndex, let tabIndex, let paneIndex):
-      do {
-        return try resolveCreatePaneTarget(
-          .pane(
-            windowIndex: windowIndex,
-            spaceIndex: spaceIndex,
-            tabIndex: tabIndex,
-            paneIndex: paneIndex
-          )
-        )
-      } catch TerminalCreatePaneError.spaceNotFound(let resolvedWindowIndex, let resolvedSpaceIndex) {
-        throw TerminalControlError.spaceNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex
-        )
-      } catch TerminalCreatePaneError.tabNotFound(
-        let resolvedWindowIndex,
-        let resolvedSpaceIndex,
-        let resolvedTabIndex
-      ) {
-        throw TerminalControlError.tabNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex,
-          tabIndex: resolvedTabIndex
-        )
-      } catch TerminalCreatePaneError.paneNotFound(
-        let resolvedWindowIndex,
-        let resolvedSpaceIndex,
-        let resolvedTabIndex,
-        let resolvedPaneIndex
-      ) {
-        throw TerminalControlError.paneNotFound(
-          windowIndex: resolvedWindowIndex,
-          spaceIndex: resolvedSpaceIndex,
-          tabIndex: resolvedTabIndex,
-          paneIndex: resolvedPaneIndex
-        )
-      } catch TerminalCreatePaneError.windowNotFound(let resolvedWindowIndex) {
-        throw TerminalControlError.windowNotFound(resolvedWindowIndex)
-      } catch {
-        throw TerminalControlError.paneNotFound(
-          windowIndex: windowIndex,
-          spaceIndex: spaceIndex,
-          tabIndex: tabIndex,
-          paneIndex: paneIndex
-        )
-      }
-    }
-  }
-
-  private func resolvedNavigationSpaceID(_ request: TerminalSpaceNavigationRequest) throws
-    -> TerminalSpaceID
-  {
-    if let windowIndex = request.windowIndex, windowIndex != 1 {
-      throw TerminalControlError.windowNotFound(windowIndex)
-    }
-    if let contextPaneID = request.contextPaneID {
-      guard let tabID = tabID(containing: contextPaneID), let space = spaceManager.space(for: tabID)
-      else {
-        throw TerminalControlError.contextPaneNotFound
-      }
-      return space.id
-    }
-    guard let selectedSpaceID else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    return selectedSpaceID
-  }
-
-  private func resolvedNavigationSpaceID(_ request: TerminalTabNavigationRequest) throws
-    -> TerminalSpaceID
-  {
-    if let windowIndex = request.windowIndex, windowIndex != 1 {
-      throw TerminalControlError.windowNotFound(windowIndex)
-    }
-    if let contextPaneID = request.contextPaneID {
-      guard let tabID = tabID(containing: contextPaneID), let space = spaceManager.space(for: tabID)
-      else {
-        throw TerminalControlError.contextPaneNotFound
-      }
-      return space.id
-    }
-    if let spaceIndex = request.spaceIndex {
-      do {
-        return try resolveSpace(windowIndex: request.windowIndex ?? 1, spaceIndex: spaceIndex).id
-      } catch TerminalCreatePaneError.spaceNotFound(let windowIndex, let resolvedSpaceIndex) {
-        throw TerminalControlError.spaceNotFound(
-          windowIndex: windowIndex, spaceIndex: resolvedSpaceIndex)
-      } catch TerminalCreatePaneError.windowNotFound(let windowIndex) {
-        throw TerminalControlError.windowNotFound(windowIndex)
-      } catch {
-        throw TerminalControlError.spaceNotFound(
-          windowIndex: request.windowIndex ?? 1, spaceIndex: spaceIndex)
-      }
-    }
-    guard let selectedSpaceID else {
-      throw TerminalControlError.lastTabNotFound
-    }
-    return selectedSpaceID
-  }
-
-  private func spaceTarget(for spaceID: TerminalSpaceID) throws -> SupatermSpaceTarget {
-    guard let space = spaces.first(where: { $0.id == spaceID }) else {
-      throw TerminalControlError.spaceNotFound(windowIndex: 1, spaceIndex: 1)
-    }
-    guard let spaceIndex = spaceManager.spaceIndex(for: spaceID) else {
-      throw TerminalControlError.spaceNotFound(windowIndex: 1, spaceIndex: 1)
-    }
-    return .init(
-      windowIndex: 1,
-      spaceIndex: spaceIndex,
-      spaceID: spaceID.rawValue,
-      name: space.name
-    )
-  }
-
-  private func tabTarget(for tabID: TerminalTabID) throws -> SupatermTabTarget {
-    guard let space = spaceManager.space(for: tabID) else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    guard let spaceIndex = spaceManager.spaceIndex(for: space.id) else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    let tabs = spaceManager.tabs(in: space.id)
-    guard let tabIndex = tabs.firstIndex(where: { $0.id == tabID }) else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: spaceIndex, tabIndex: 1)
-    }
-    let tab = tabs[tabIndex]
-    return .init(
-      windowIndex: 1,
-      spaceIndex: spaceIndex,
-      spaceID: space.id.rawValue,
-      tabIndex: tabIndex + 1,
-      tabID: tabID.rawValue,
-      title: tab.title
-    )
-  }
-
-  private func paneTarget(
-    spaceID: TerminalSpaceID,
-    tabID: TerminalTabID,
-    surfaceID: UUID,
-    tree: SplitTree<GhosttySurfaceView>
-  ) throws -> SupatermPaneTarget {
-    let location = try resolvedPaneLocation(
-      spaceID: spaceID,
-      tabID: tabID,
-      surfaceID: surfaceID,
-      tree: tree
-    )
-    return .init(
-      windowIndex: 1,
-      spaceIndex: location.spaceIndex,
-      spaceID: spaceID.rawValue,
-      tabIndex: location.tabIndex,
-      tabID: tabID.rawValue,
-      paneIndex: location.paneIndex,
-      paneID: surfaceID
-    )
-  }
-
-  private func resolvedFocusedSurface(
-    in tabID: TerminalTabID
-  ) -> (tree: SplitTree<GhosttySurfaceView>, surface: GhosttySurfaceView)? {
-    guard let tree = trees[tabID] else { return nil }
-    if let focusedSurfaceID = focusedSurfaceIDByTab[tabID], let surface = surfaces[focusedSurfaceID] {
-      return (tree, surface)
-    }
-    guard let surface = tree.root?.leftmostLeaf() else { return nil }
-    return (tree, surface)
-  }
-
-  private func focusPaneResult(
-    spaceID: TerminalSpaceID,
-    tabID: TerminalTabID,
-    surfaceID: UUID,
-    tree: SplitTree<GhosttySurfaceView>
-  ) throws -> SupatermFocusPaneResult {
-    let target = try paneTarget(
-      spaceID: spaceID,
-      tabID: tabID,
-      surfaceID: surfaceID,
-      tree: tree
-    )
-    let activity = Self.surfaceActivity(
-      isSelectedTab: selectedTabID == tabID,
-      windowIsVisible: windowActivity.isVisible,
-      windowIsKey: windowActivity.isKeyWindow,
-      focusedSurfaceID: focusedSurfaceIDByTab[tabID],
-      surfaceID: surfaceID
-    )
-    return .init(
-      isFocused: activity.isFocused,
-      isSelectedTab: selectedTabID == tabID,
-      target: target
-    )
-  }
-
-  private func selectTabResult(for tabID: TerminalTabID) throws -> SupatermSelectTabResult {
-    guard let space = spaceManager.space(for: tabID) else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    guard let resolvedSurface = resolvedFocusedSurface(in: tabID) else {
-      throw TerminalControlError.tabNotFound(windowIndex: 1, spaceIndex: 1, tabIndex: 1)
-    }
-    let target = try tabTarget(for: tabID)
-    let paneTarget = try paneTarget(
-      spaceID: space.id,
-      tabID: tabID,
-      surfaceID: resolvedSurface.surface.id,
-      tree: resolvedSurface.tree
-    )
-    let activity = Self.surfaceActivity(
-      isSelectedTab: selectedTabID == tabID,
-      windowIsVisible: windowActivity.isVisible,
-      windowIsKey: windowActivity.isKeyWindow,
-      focusedSurfaceID: focusedSurfaceIDByTab[tabID],
-      surfaceID: resolvedSurface.surface.id
-    )
-    return .init(
-      isFocused: activity.isFocused,
-      isSelectedSpace: selectedSpaceID == space.id,
-      isSelectedTab: selectedTabID == tabID,
-      isTitleLocked: spaceManager.tab(for: tabID)?.isTitleLocked == true,
-      paneIndex: paneTarget.paneIndex,
-      paneID: paneTarget.paneID,
-      target: target
-    )
-  }
-
-  private func selectSpaceResult(for spaceID: TerminalSpaceID) throws -> SupatermSelectSpaceResult {
-    guard
-      let tabID = spaceManager.selectedTabID(in: spaceID)
-        ?? spaceManager.tabs(in: spaceID).first?.id
-    else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    let target = try spaceTarget(for: spaceID)
-    let tabTarget = try self.tabTarget(for: tabID)
-    guard let resolvedSurface = resolvedFocusedSurface(in: tabID) else {
-      throw TerminalControlError.lastSpaceNotFound
-    }
-    let paneTarget = try paneTarget(
-      spaceID: spaceID,
-      tabID: tabID,
-      surfaceID: resolvedSurface.surface.id,
-      tree: resolvedSurface.tree
-    )
-    let activity = Self.surfaceActivity(
-      isSelectedTab: selectedTabID == tabID,
-      windowIsVisible: windowActivity.isVisible,
-      windowIsKey: windowActivity.isKeyWindow,
-      focusedSurfaceID: focusedSurfaceIDByTab[tabID],
-      surfaceID: resolvedSurface.surface.id
-    )
-    return .init(
-      isFocused: activity.isFocused,
-      isSelectedSpace: selectedSpaceID == spaceID,
-      isSelectedTab: selectedTabID == tabID,
-      paneIndex: paneTarget.paneIndex,
-      paneID: paneTarget.paneID,
-      tabIndex: tabTarget.tabIndex,
-      tabID: tabTarget.tabID,
-      target: target
-    )
-  }
-
-  private func notify(
-    _ request: TerminalNotifyRequest,
-    origin: NotificationOrigin
-  ) throws -> SupatermNotifyResult {
-    let resolvedTarget = try resolveNotifyTarget(request.target)
-    let paneLocation = try resolvedPaneLocation(
-      spaceID: resolvedTarget.spaceID,
-      tabID: resolvedTarget.tabID,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      tree: resolvedTarget.tree
-    )
-    let selectionState = Self.newPaneSelectionState(
-      selectedTabID: spaceManager.selectedTabID,
-      targetTabID: resolvedTarget.tabID,
-      windowActivity: windowActivity,
-      focusedSurfaceID: focusedSurfaceIDByTab[resolvedTarget.tabID],
-      surfaceID: resolvedTarget.anchorSurface.id
-    )
-    let attentionState: SupatermNotificationAttentionState = .unread
-    let desktopNotificationDisposition = resolvedDesktopNotificationDisposition(
-      allowDesktopNotificationWhenAgentActive: request.allowDesktopNotificationWhenAgentActive,
-      isFocused: selectionState.isFocused,
-      tabID: resolvedTarget.tabID
-    )
-    let resolvedTitle = resolvedNotificationTitle(
-      request.title,
-      for: resolvedTarget.tabID
-    )
-    let createdAt = Date()
-    coalesceStructuredNotificationIfNeeded(
-      body: request.body,
-      origin: origin,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      title: resolvedTitle
-    )
-    paneNotifications[resolvedTarget.anchorSurface.id, default: []].append(
-      PaneNotification(
-        attentionState: attentionState,
-        body: request.body,
-        createdAt: createdAt,
-        subtitle: request.subtitle,
-        title: resolvedTitle,
-        origin: origin
-      ))
-    updateRecentStructuredNotificationIfNeeded(
-      body: request.body,
-      createdAt: createdAt,
-      origin: origin,
-      surfaceID: resolvedTarget.anchorSurface.id,
-      title: resolvedTitle
-    )
-
-    return .init(
-      attentionState: attentionState,
-      desktopNotificationDisposition: desktopNotificationDisposition,
-      resolvedTitle: resolvedTitle,
-      windowIndex: 1,
-      spaceIndex: paneLocation.spaceIndex,
-      spaceID: resolvedTarget.spaceID.rawValue,
-      tabIndex: paneLocation.tabIndex,
-      tabID: resolvedTarget.tabID.rawValue,
-      paneIndex: paneLocation.paneIndex,
-      paneID: resolvedTarget.anchorSurface.id
-    )
-  }
-
-  private func handleDesktopNotification(
-    body: String,
-    surfaceID: UUID,
-    title: String
-  ) {
-    let subtitle = ""
-    guard !shouldSuppressDesktopNotification(body: body, surfaceID: surfaceID, title: title) else {
-      return
-    }
-    guard
-      let result = try? notify(
-        .init(
-          body: body,
-          subtitle: subtitle,
-          target: .contextPane(surfaceID),
-          title: Self.trimmedNonEmpty(title)
-        ),
-        origin: .terminalDesktop
-      )
-    else {
-      return
-    }
-    emit(
-      .notificationReceived(
-        .init(
-          attentionState: result.attentionState,
-          body: body,
-          desktopNotificationDisposition: result.desktopNotificationDisposition,
-          resolvedTitle: result.resolvedTitle,
-          sourceSurfaceID: surfaceID,
-          subtitle: subtitle
-        )
-      )
-    )
-  }
-
-  func handleDirectInteraction(on surfaceID: UUID) {
-    clearNotificationAttention(for: surfaceID)
-  }
-
   func splitTree(
     for tabID: TerminalTabID,
     inheritingFromSurfaceID: UUID? = nil,
@@ -2253,7 +827,7 @@ final class TerminalHostState {
     return tree
   }
 
-  private func performSplitAction(_ action: GhosttySplitAction, for surfaceID: UUID) -> Bool {
+  func performSplitAction(_ action: GhosttySplitAction, for surfaceID: UUID) -> Bool {
     guard let tabID = tabID(containing: surfaceID), var tree = trees[tabID] else {
       return false
     }
@@ -2327,7 +901,7 @@ final class TerminalHostState {
     }
   }
 
-  private func performSplitOperation(
+  func performSplitOperation(
     _ operation: TerminalSplitTreeView.Operation, in tabID: TerminalTabID
   ) {
     guard var tree = trees[tabID] else { return }
@@ -2381,7 +955,7 @@ final class TerminalHostState {
     return SurfaceActivity(isVisible: isVisible, isFocused: isFocused)
   }
 
-  private func performCloseTab(_ tabID: TerminalTabID) {
+  func performCloseTab(_ tabID: TerminalTabID) {
     guard let space = spaceManager.space(for: tabID) else { return }
     guard let tabManager = spaceManager.tabManager(for: space.id) else { return }
 
@@ -2409,13 +983,13 @@ final class TerminalHostState {
     sessionDidChange()
   }
 
-  private func performCloseTabs(_ tabIDs: [TerminalTabID]) {
+  func performCloseTabs(_ tabIDs: [TerminalTabID]) {
     for tabID in tabIDs {
       performCloseTab(tabID)
     }
   }
 
-  private func performCloseSurface(_ surfaceID: UUID) {
+  func performCloseSurface(_ surfaceID: UUID) {
     guard let tabID = tabID(containing: surfaceID), let tree = trees[tabID] else { return }
     guard let node = tree.find(id: surfaceID), let surface = surfaces[surfaceID] else { return }
 
@@ -2469,7 +1043,7 @@ final class TerminalHostState {
     sessionDidChange()
   }
 
-  private func createSurface(
+  func createSurface(
     tabID: TerminalTabID,
     command: String?,
     initialInput: String?,
@@ -2497,7 +1071,7 @@ final class TerminalHostState {
     return view
   }
 
-  private func configureBridgeCallbacks(
+  func configureBridgeCallbacks(
     for view: GhosttySurfaceView,
     tabID: TerminalTabID
   ) {
@@ -2572,7 +1146,7 @@ final class TerminalHostState {
     }
   }
 
-  private func configureSurfaceCallbacks(
+  func configureSurfaceCallbacks(
     for view: GhosttySurfaceView,
     tabID: TerminalTabID
   ) {
@@ -2591,12 +1165,12 @@ final class TerminalHostState {
     }
   }
 
-  private struct InheritedSurfaceConfig: Equatable {
+  struct InheritedSurfaceConfig: Equatable {
     let workingDirectory: URL?
     let fontSize: Float32?
   }
 
-  private func inheritedSurfaceConfig(
+  func inheritedSurfaceConfig(
     fromSurfaceID surfaceID: UUID?,
     context: ghostty_surface_context_e
   ) -> InheritedSurfaceConfig {
@@ -2618,12 +1192,12 @@ final class TerminalHostState {
     )
   }
 
-  private func currentFocusedSurfaceID() -> UUID? {
+  func currentFocusedSurfaceID() -> UUID? {
     guard let selectedTabID = spaceManager.selectedTabID else { return nil }
     return focusedSurfaceIDByTab[selectedTabID]
   }
 
-  private func inheritedSurfaceID(in spaceID: TerminalSpaceID) -> UUID? {
+  func inheritedSurfaceID(in spaceID: TerminalSpaceID) -> UUID? {
     if let selectedTabID = spaceManager.selectedTabID(in: spaceID) {
       if let focusedSurfaceID = focusedSurfaceIDByTab[selectedTabID],
         surfaces[focusedSurfaceID] != nil
@@ -2647,32 +1221,32 @@ final class TerminalHostState {
     return nil
   }
 
-  private struct ResolvedCreateTabTarget {
+  struct ResolvedCreateTabTarget {
     let anchorTabID: TerminalTabID?
     let inheritedSurfaceID: UUID?
     let space: TerminalSpaceItem
   }
 
-  private struct ResolvedLocalCreateTabTarget {
+  struct ResolvedLocalCreateTabTarget {
     let anchorTabID: TerminalTabID?
     let inheritedSurfaceID: UUID?
     let spaceID: TerminalSpaceID
   }
 
-  private struct ResolvedCreatePaneTarget {
+  struct ResolvedCreatePaneTarget {
     let anchorSurface: GhosttySurfaceView
     let spaceID: TerminalSpaceID
     let tabID: TerminalTabID
     let tree: SplitTree<GhosttySurfaceView>
   }
 
-  private struct ResolvedCreatePaneTab {
+  struct ResolvedCreatePaneTab {
     let space: TerminalSpaceItem
     let tabID: TerminalTabID
     let tree: SplitTree<GhosttySurfaceView>
   }
 
-  private func resolveCreateTabTarget(
+  func resolveCreateTabTarget(
     _ target: TerminalCreateTabRequest.Target
   ) throws -> ResolvedCreateTabTarget {
     switch target {
@@ -2705,7 +1279,7 @@ final class TerminalHostState {
     }
   }
 
-  private func resolveLocalCreateTabTarget(
+  func resolveLocalCreateTabTarget(
     inheritingFromSurfaceID: UUID?
   ) -> ResolvedLocalCreateTabTarget? {
     if let inheritingFromSurfaceID,
@@ -2730,7 +1304,7 @@ final class TerminalHostState {
     )
   }
 
-  private func resolvedNewTabInsertion(
+  func resolvedNewTabInsertion(
     anchorTabID: TerminalTabID?
   ) -> TerminalTabManager.Insertion {
     switch supatermSettings.newTabPosition {
@@ -2744,13 +1318,13 @@ final class TerminalHostState {
     }
   }
 
-  private struct ResolvedPaneLocation {
+  struct ResolvedPaneLocation {
     let paneIndex: Int
     let spaceIndex: Int
     let tabIndex: Int
   }
 
-  private func resolveCreatePaneTarget(
+  func resolveCreatePaneTarget(
     _ target: TerminalCreatePaneRequest.Target
   ) throws -> ResolvedCreatePaneTarget {
     switch target {
@@ -2811,7 +1385,7 @@ final class TerminalHostState {
     }
   }
 
-  private func resolveNotifyTarget(
+  func resolveNotifyTarget(
     _ target: TerminalNotifyRequest.Target
   ) throws -> ResolvedCreatePaneTarget {
     switch target {
@@ -2830,7 +1404,7 @@ final class TerminalHostState {
     }
   }
 
-  private func resolveSpace(
+  func resolveSpace(
     windowIndex: Int,
     spaceIndex: Int
   ) throws -> TerminalSpaceItem {
@@ -2843,7 +1417,7 @@ final class TerminalHostState {
     return space
   }
 
-  private func resolveTab(
+  func resolveTab(
     windowIndex: Int,
     spaceIndex: Int,
     tabIndex: Int
@@ -2867,7 +1441,7 @@ final class TerminalHostState {
     return .init(space: space, tabID: tabID, tree: tree)
   }
 
-  private func resolvedPaneLocation(
+  func resolvedPaneLocation(
     spaceID: TerminalSpaceID,
     tabID: TerminalTabID,
     surfaceID: UUID,
@@ -2888,14 +1462,14 @@ final class TerminalHostState {
     )
   }
 
-  private func updateTabTitle(for tabID: TerminalTabID) {
+  func updateTabTitle(for tabID: TerminalTabID) {
     let resolvedTitle = currentTabTitle(for: tabID)
     spaceManager.space(for: tabID)
       .flatMap { spaceManager.tabManager(for: $0.id) }?
       .updateTitle(tabID, title: resolvedTitle)
   }
 
-  private func focusSurface(in tabID: TerminalTabID) {
+  func focusSurface(in tabID: TerminalTabID) {
     if let unreadSurfaceID = latestUnreadNotifiedSurfaceID(in: tabID),
       let surface = surfaces[unreadSurfaceID]
     {
@@ -2912,7 +1486,7 @@ final class TerminalHostState {
     }
   }
 
-  private func applyFocusedSurface(
+  func applyFocusedSurface(
     _ surfaceID: UUID,
     in tabID: TerminalTabID
   ) {
@@ -2923,7 +1497,7 @@ final class TerminalHostState {
     focusedSurfaceIDByTab[tabID] = surfaceID
   }
 
-  private func focusSurface(_ surface: GhosttySurfaceView, in tabID: TerminalTabID) {
+  func focusSurface(_ surface: GhosttySurfaceView, in tabID: TerminalTabID) {
     let previousSurface = focusedSurfaceIDByTab[tabID].flatMap { surfaces[$0] }
     applyFocusedSurface(surface.id, in: tabID)
     updateTabTitle(for: tabID)
@@ -3000,7 +1574,7 @@ final class TerminalHostState {
     )
   }
 
-  private func removeTree(for tabID: TerminalTabID) {
+  func removeTree(for tabID: TerminalTabID) {
     guard let tree = trees.removeValue(forKey: tabID) else { return }
     for surface in tree.leaves() {
       paneNotifications.removeValue(forKey: surface.id)
@@ -3016,7 +1590,7 @@ final class TerminalHostState {
     previousSelectedTabIDBySpace = previousSelectedTabIDBySpace.filter { $0.value != tabID }
   }
 
-  private func notifications(for tabID: TerminalTabID) -> [UUID: [PaneNotification]] {
+  func notifications(for tabID: TerminalTabID) -> [UUID: [PaneNotification]] {
     guard let tree = trees[tabID] else { return [:] }
     return Dictionary(
       uniqueKeysWithValues: tree.leaves().compactMap { surface in
@@ -3025,7 +1599,7 @@ final class TerminalHostState {
     )
   }
 
-  private func latestUnreadNotifiedSurfaceID(in tabID: TerminalTabID) -> UUID? {
+  func latestUnreadNotifiedSurfaceID(in tabID: TerminalTabID) -> UUID? {
     notifications(for: tabID)
       .compactMap { surfaceID, notifications -> (UUID, PaneNotification)? in
         guard
@@ -3043,11 +1617,11 @@ final class TerminalHostState {
       .0
   }
 
-  private func latestNotification(for tabID: TerminalTabID) -> PaneNotification? {
+  func latestNotification(for tabID: TerminalTabID) -> PaneNotification? {
     Self.latestNotification(in: notifications(for: tabID).values.flatMap { $0 })
   }
 
-  private func clearUnreadOnFocusedSurfaceIfNeeded() {
+  func clearUnreadOnFocusedSurfaceIfNeeded() {
     guard
       let selectedTabID = spaceManager.selectedTabID,
       let surfaceID = focusedSurfaceIDByTab[selectedTabID]
@@ -3057,7 +1631,7 @@ final class TerminalHostState {
     clearNotificationAttention(for: surfaceID)
   }
 
-  private func clearNotificationAttention(for surfaceID: UUID) {
+  func clearNotificationAttention(for surfaceID: UUID) {
     guard let tabID = tabID(containing: surfaceID) else { return }
     let activity = Self.surfaceActivity(
       isSelectedTab: tabID == spaceManager.selectedTabID,
@@ -3077,7 +1651,7 @@ final class TerminalHostState {
     paneNotifications[surfaceID] = updatedNotifications
   }
 
-  private func updateRunningState(for tabID: TerminalTabID) {
+  func updateRunningState(for tabID: TerminalTabID) {
     guard let tree = trees[tabID] else { return }
     let isRunning = tree.leaves().contains { surface in
       Self.isRunning(progressState: surface.bridge.state.progressState)
@@ -3087,7 +1661,7 @@ final class TerminalHostState {
       .updateDirty(tabID, isDirty: isRunning)
   }
 
-  private func tabNeedsCloseConfirmation(_ tabID: TerminalTabID) -> Bool {
+  func tabNeedsCloseConfirmation(_ tabID: TerminalTabID) -> Bool {
     guard let tree = trees[tabID] else { return false }
     return tree.leaves().contains(where: \.needsCloseConfirmation)
   }
@@ -3101,23 +1675,23 @@ final class TerminalHostState {
     }
   }
 
-  private func surfaceNeedsCloseConfirmation(_ surfaceID: UUID) -> Bool {
+  func surfaceNeedsCloseConfirmation(_ surfaceID: UUID) -> Bool {
     surfaces[surfaceID]?.needsCloseConfirmation ?? false
   }
 
-  private func tabID(containing surfaceID: UUID) -> TerminalTabID? {
+  func tabID(containing surfaceID: UUID) -> TerminalTabID? {
     for (tabID, tree) in trees where tree.find(id: surfaceID) != nil {
       return tabID
     }
     return nil
   }
 
-  private func emitFocusChangedIfNeeded(_ surfaceID: UUID) {
+  func emitFocusChangedIfNeeded(_ surfaceID: UUID) {
     guard surfaceID != lastEmittedFocusSurfaceID else { return }
     lastEmittedFocusSurfaceID = surfaceID
   }
 
-  private func emit(_ event: TerminalClient.Event) {
+  func emit(_ event: TerminalClient.Event) {
     guard let eventContinuation else {
       pendingEvents.append(event)
       return
@@ -3125,7 +1699,7 @@ final class TerminalHostState {
     eventContinuation.yield(event)
   }
 
-  private func mapGotoTabTarget(_ target: ghostty_action_goto_tab_e) -> TerminalGotoTabTarget? {
+  func mapGotoTabTarget(_ target: ghostty_action_goto_tab_e) -> TerminalGotoTabTarget? {
     let raw = Int(target.rawValue)
     if raw <= 0 {
       switch raw {
@@ -3142,7 +1716,7 @@ final class TerminalHostState {
     return .index(raw)
   }
 
-  private func nextTabIndex(in spaceID: TerminalSpaceID) -> Int {
+  func nextTabIndex(in spaceID: TerminalSpaceID) -> Int {
     var maxIndex = 0
     for tab in spaceManager.tabs(in: spaceID) {
       guard tab.title.hasPrefix("Terminal ") else { continue }
@@ -3153,11 +1727,11 @@ final class TerminalHostState {
     return maxIndex + 1
   }
 
-  private func fallbackTitle(for tabID: TerminalTabID) -> String {
+  func fallbackTitle(for tabID: TerminalTabID) -> String {
     spaceManager.tab(for: tabID)?.defaultTitle ?? "Terminal"
   }
 
-  private func setLockedTabTitle(_ title: String?, for tabID: TerminalTabID) {
+  func setLockedTabTitle(_ title: String?, for tabID: TerminalTabID) {
     spaceManager.space(for: tabID)
       .flatMap { spaceManager.tabManager(for: $0.id) }?
       .setLockedTitle(tabID, title: title)
@@ -3170,7 +1744,7 @@ final class TerminalHostState {
     promptTabTitle(for: tabID, using: view)
   }
 
-  private func promptTabTitle(for tabID: TerminalTabID, using view: GhosttySurfaceView) {
+  func promptTabTitle(for tabID: TerminalTabID, using view: GhosttySurfaceView) {
     view.promptTitle(
       messageText: "Change Tab Title",
       initialValue: currentTabTitle(for: tabID)
@@ -3179,7 +1753,7 @@ final class TerminalHostState {
     }
   }
 
-  private func copyTitleToClipboard(for surfaceID: UUID) -> Bool {
+  func copyTitleToClipboard(for surfaceID: UUID) -> Bool {
     guard let surface = surfaces[surfaceID] else { return false }
     guard let title = surface.effectiveTitle(), !title.isEmpty else { return false }
     let pasteboard = NSPasteboard.general
@@ -3187,7 +1761,7 @@ final class TerminalHostState {
     return pasteboard.setString(title, forType: .string)
   }
 
-  private func currentTabTitle(for tabID: TerminalTabID) -> String {
+  func currentTabTitle(for tabID: TerminalTabID) -> String {
     if let title = lockedTabTitle(for: tabID) {
       return title
     }
@@ -3203,19 +1777,19 @@ final class TerminalHostState {
     )
   }
 
-  private func lockedTabTitle(for tabID: TerminalTabID) -> String? {
+  func lockedTabTitle(for tabID: TerminalTabID) -> String? {
     guard let tab = spaceManager.tab(for: tabID), tab.isTitleLocked else { return nil }
     return tab.title
   }
 
-  private func resolvedNotificationTitle(
+  func resolvedNotificationTitle(
     _ title: String?,
     for tabID: TerminalTabID
   ) -> String {
     Self.trimmedNonEmpty(title) ?? currentTabTitle(for: tabID)
   }
 
-  private func resolvedDesktopNotificationDisposition(
+  func resolvedDesktopNotificationDisposition(
     allowDesktopNotificationWhenAgentActive: Bool,
     isFocused: Bool,
     tabID: TerminalTabID
@@ -3229,7 +1803,7 @@ final class TerminalHostState {
     return .deliver
   }
 
-  private func hasActiveAgentAttention(for tabID: TerminalTabID) -> Bool {
+  func hasActiveAgentAttention(for tabID: TerminalTabID) -> Bool {
     switch agentActivityByTab[tabID]?.phase {
     case .some(.needsInput), .some(.running):
       return true
@@ -3238,170 +1812,14 @@ final class TerminalHostState {
     }
   }
 
-  private func titleSurface(for tabID: TerminalTabID) -> GhosttySurfaceView? {
+  func titleSurface(for tabID: TerminalTabID) -> GhosttySurfaceView? {
     if let focusedSurfaceID = focusedSurfaceIDByTab[tabID] {
       return surfaces[focusedSurfaceID]
     }
     return trees[tabID]?.root?.leftmostLeaf()
   }
 
-  private func restorationTabSession(for tab: TerminalTabItem) -> TerminalTabSession? {
-    guard let tree = trees[tab.id], let root = tree.root.map(restorationNode(for:)) else {
-      return nil
-    }
-    let focusedPaneIndex =
-      focusedSurfaceIDByTab[tab.id].flatMap { focusedPaneID in
-        tree.leaves().firstIndex(where: { $0.id == focusedPaneID })
-      }
-      ?? 0
-    return TerminalTabSession(
-      isPinned: tab.isPinned,
-      lockedTitle: lockedTabTitle(for: tab.id),
-      focusedPaneIndex: focusedPaneIndex,
-      root: root
-    )
-  }
-
-  private func restorationNode(
-    for node: SplitTree<GhosttySurfaceView>.Node
-  ) -> TerminalPaneNodeSession {
-    switch node {
-    case .leaf(let surface):
-      return .leaf(
-        TerminalPaneLeafSession(
-          workingDirectoryPath: workingDirectoryPath(for: surface),
-          titleOverride: surface.bridge.state.titleOverride
-        )
-      )
-    case .split(let split):
-      return .split(
-        TerminalPaneSplitSession(
-          direction: mapSessionSplitDirection(split.direction),
-          ratio: split.ratio,
-          left: restorationNode(for: split.left),
-          right: restorationNode(for: split.right)
-        )
-      )
-    }
-  }
-
-  private func restoredTabItem(
-    for session: TerminalTabSession,
-    at index: Int
-  ) -> TerminalTabItem {
-    TerminalTabItem(
-      title: session.lockedTitle ?? restoredTabTitle(at: index),
-      icon: "terminal",
-      isPinned: session.isPinned,
-      isTitleLocked: session.lockedTitle != nil
-    )
-  }
-
-  private func restoredTabTitle(at index: Int) -> String {
-    index == 0 ? "Terminal" : "Terminal \(index + 1)"
-  }
-
-  private func restoreTabSession(
-    _ session: TerminalTabSession,
-    tabID: TerminalTabID,
-    in spaceID: TerminalSpaceID
-  ) {
-    let context: ghostty_surface_context_e =
-      spaceManager.tabs(in: spaceID).first?.id == tabID
-      ? GHOSTTY_SURFACE_CONTEXT_WINDOW
-      : GHOSTTY_SURFACE_CONTEXT_TAB
-    let restoredRoot = restoreNode(
-      session.root,
-      in: tabID,
-      context: context
-    )
-    trees[tabID] = SplitTree(root: restoredRoot, zoomed: nil)
-    let leaves = restoredRoot.leaves()
-    let focusedPaneIndex =
-      leaves.indices.contains(session.focusedPaneIndex)
-      ? session.focusedPaneIndex
-      : 0
-    applyFocusedSurface(leaves[focusedPaneIndex].id, in: tabID)
-    updateRunningState(for: tabID)
-    updateTabTitle(for: tabID)
-  }
-
-  private func restoreNode(
-    _ node: TerminalPaneNodeSession,
-    in tabID: TerminalTabID,
-    context: ghostty_surface_context_e
-  ) -> SplitTree<GhosttySurfaceView>.Node {
-    switch node {
-    case .leaf(let leaf):
-      let surface = createSurface(
-        tabID: tabID,
-        command: nil,
-        initialInput: nil,
-        inheritingFromSurfaceID: nil,
-        workingDirectory: existingWorkingDirectoryURL(for: leaf.workingDirectoryPath),
-        context: context
-      )
-      surface.bridge.state.titleOverride = leaf.titleOverride
-      return .leaf(view: surface)
-
-    case .split(let split):
-      let left = restoreNode(split.left, in: tabID, context: GHOSTTY_SURFACE_CONTEXT_SPLIT)
-      let right = restoreNode(split.right, in: tabID, context: GHOSTTY_SURFACE_CONTEXT_SPLIT)
-      return .split(
-        .init(
-          direction: mapSplitDirection(split.direction),
-          ratio: split.ratio,
-          left: left,
-          right: right
-        )
-      )
-    }
-  }
-
-  private func clearSessionState() {
-    let existingTabIDs = spaces.flatMap { spaceManager.tabs(in: $0.id).map(\.id) }
-    removeTrees(for: existingTabIDs)
-    for space in spaces {
-      _ = spaceManager.restoreTabs([], selectedTabID: nil, in: space.id)
-    }
-    previousFocusedSurfaceIDByTab.removeAll()
-    previousSelectedTabIDBySpace.removeAll()
-    previousSelectedSpaceID = nil
-    lastEmittedFocusSurfaceID = nil
-  }
-
-  private func sessionDidChange() {
-    guard suppressesSessionChanges == 0 else { return }
-    onSessionChange()
-  }
-
-  private func withSessionChangesSuppressed<Result>(
-    _ body: () -> Result
-  ) -> Result {
-    suppressesSessionChanges += 1
-    defer {
-      suppressesSessionChanges -= 1
-    }
-    return body()
-  }
-
-  private func workingDirectoryPath(for surface: GhosttySurfaceView) -> String? {
-    guard let path = Self.trimmedNonEmpty(surface.bridge.state.pwd) else { return nil }
-    return GhosttySurfaceView.normalizedWorkingDirectoryPath(path)
-  }
-
-  private func existingWorkingDirectoryURL(for path: String?) -> URL? {
-    guard let path = Self.trimmedNonEmpty(path) else { return nil }
-    let normalizedPath = GhosttySurfaceView.normalizedWorkingDirectoryPath(path)
-    var isDirectory = ObjCBool(false)
-    guard FileManager.default.fileExists(atPath: normalizedPath, isDirectory: &isDirectory) else {
-      return nil
-    }
-    guard isDirectory.boolValue else { return nil }
-    return URL(fileURLWithPath: normalizedPath, isDirectory: true)
-  }
-
-  private func observeSpaceCatalog() {
+  func observeSpaceCatalog() {
     spaceCatalogObservationTask?.cancel()
     spaceCatalogObservationTask = Task { @MainActor [weak self] in
       let observations = Observations { [weak self] in
@@ -3414,7 +1832,7 @@ final class TerminalHostState {
     }
   }
 
-  private func applyObservedSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
+  func applyObservedSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
     let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
     guard resolvedSpaceCatalog != lastAppliedSpaceCatalog else { return }
 
@@ -3433,7 +1851,7 @@ final class TerminalHostState {
   }
 
   @discardableResult
-  private func writeSpaceCatalog(
+  func writeSpaceCatalog(
     _ spaceCatalog: TerminalSpaceCatalog
   ) -> TerminalSpaceManager.SpaceCatalogDiff {
     let resolvedSpaceCatalog = TerminalSpaceCatalog.sanitized(spaceCatalog)
@@ -3445,7 +1863,7 @@ final class TerminalHostState {
     return diff
   }
 
-  private func persistDefaultSelectedSpaceID(_ spaceID: TerminalSpaceID) {
+  func persistDefaultSelectedSpaceID(_ spaceID: TerminalSpaceID) {
     guard spaceCatalog.defaultSelectedSpaceID != spaceID else { return }
 
     var updatedSpaceCatalog = spaceCatalog
@@ -3454,11 +1872,11 @@ final class TerminalHostState {
     lastAppliedSpaceCatalog = updatedSpaceCatalog
   }
 
-  private func replaceSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
+  func replaceSpaceCatalog(_ spaceCatalog: TerminalSpaceCatalog) {
     $spaceCatalog.withLock { $0 = spaceCatalog }
   }
 
-  private func nextSelectedSpaceID(afterDeleting spaceID: TerminalSpaceID) -> TerminalSpaceID {
+  func nextSelectedSpaceID(afterDeleting spaceID: TerminalSpaceID) -> TerminalSpaceID {
     let remainingSpaces = spaces.filter { $0.id != spaceID }
     precondition(!remainingSpaces.isEmpty)
 
@@ -3479,7 +1897,7 @@ final class TerminalHostState {
     return remainingSpaces[0].id
   }
 
-  private func finalizeSpaceSelectionChange() {
+  func finalizeSpaceSelectionChange() {
     guard managesTerminalSurfaces else {
       lastEmittedFocusSurfaceID = nil
       return
@@ -3493,13 +1911,13 @@ final class TerminalHostState {
     syncFocus(windowActivity)
   }
 
-  private func removeTrees(for tabIDs: [TerminalTabID]) {
+  func removeTrees(for tabIDs: [TerminalTabID]) {
     for tabID in tabIDs {
       removeTree(for: tabID)
     }
   }
 
-  private func mapSplitDirection(_ direction: GhosttySplitAction.NewDirection)
+  func mapSplitDirection(_ direction: GhosttySplitAction.NewDirection)
     -> SplitTree<GhosttySurfaceView>.NewDirection
   {
     switch direction {
@@ -3514,7 +1932,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapSplitDirection(_ direction: TerminalPaneSplitDirection)
+  func mapSplitDirection(_ direction: TerminalPaneSplitDirection)
     -> SplitTree<GhosttySurfaceView>.Direction
   {
     switch direction {
@@ -3525,7 +1943,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapSessionSplitDirection(
+  func mapSessionSplitDirection(
     _ direction: SplitTree<GhosttySurfaceView>.Direction
   ) -> TerminalPaneSplitDirection {
     switch direction {
@@ -3536,7 +1954,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapPaneDirection(_ direction: SupatermPaneDirection)
+  func mapPaneDirection(_ direction: SupatermPaneDirection)
     -> SplitTree<GhosttySurfaceView>.NewDirection
   {
     switch direction {
@@ -3551,7 +1969,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapFocusDirection(_ direction: GhosttySplitAction.FocusDirection)
+  func mapFocusDirection(_ direction: GhosttySplitAction.FocusDirection)
     -> SplitTree<GhosttySurfaceView>.FocusDirection
   {
     switch direction {
@@ -3570,7 +1988,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapResizeDirection(_ direction: GhosttySplitAction.ResizeDirection)
+  func mapResizeDirection(_ direction: GhosttySplitAction.ResizeDirection)
     -> SplitTree<GhosttySurfaceView>.SpatialDirection
   {
     switch direction {
@@ -3585,7 +2003,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapResizeDirection(_ direction: SupatermResizePaneDirection)
+  func mapResizeDirection(_ direction: SupatermResizePaneDirection)
     -> SplitTree<GhosttySurfaceView>.SpatialDirection
   {
     switch direction {
@@ -3600,7 +2018,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapPaneAxis(_ axis: SupatermPaneAxis)
+  func mapPaneAxis(_ axis: SupatermPaneAxis)
     -> SplitTree<GhosttySurfaceView>.Direction
   {
     switch axis {
@@ -3611,7 +2029,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapPaneSizeUnit(_ unit: SupatermPaneSizeUnit)
+  func mapPaneSizeUnit(_ unit: SupatermPaneSizeUnit)
     -> SplitTree<GhosttySurfaceView>.SizeUnit
   {
     switch unit {
@@ -3622,7 +2040,7 @@ final class TerminalHostState {
     }
   }
 
-  private func mapDropZone(_ zone: TerminalSplitTreeView.DropZone)
+  func mapDropZone(_ zone: TerminalSplitTreeView.DropZone)
     -> SplitTree<GhosttySurfaceView>.NewDirection
   {
     switch zone {
@@ -3637,7 +2055,7 @@ final class TerminalHostState {
     }
   }
 
-  private func debugPaneSnapshot(
+  func debugPaneSnapshot(
     _ surface: GhosttySurfaceView,
     index: Int,
     isFocused: Bool
@@ -3663,513 +2081,4 @@ final class TerminalHostState {
     )
   }
 
-  static func selectedPaneDisplayTitle<Surface: NSView & Identifiable>(
-    focusedSurfaceID: UUID?,
-    in tree: SplitTree<Surface>?,
-    titleOverride: (Surface) -> String?,
-    title: (Surface) -> String?,
-    pwd: (Surface) -> String?
-  ) -> String where Surface.ID == UUID {
-    let leaves = tree?.leaves() ?? []
-    guard
-      let surface = focusedSurfaceID.flatMap({ id in leaves.first(where: { $0.id == id }) })
-        ?? leaves.first
-    else {
-      return "Pane"
-    }
-    return resolvedPaneDisplayTitle(
-      titleOverride: titleOverride(surface),
-      title: title(surface),
-      pwd: pwd(surface),
-      defaultValue: paneFallbackTitle(for: surface.id, in: tree)
-    )
-  }
-
-  static func paneWorkingDirectories<Surface: NSView & Identifiable>(
-    in tree: SplitTree<Surface>?,
-    pwd: (Surface) -> String?
-  ) -> [String] where Surface.ID == UUID {
-    var seen = Set<String>()
-    return (tree?.leaves() ?? []).compactMap { surface in
-      guard let path = trimmedNonEmpty(pwd(surface)) else { return nil }
-      let normalized = GhosttySurfaceView.normalizedWorkingDirectoryPath(path)
-      guard seen.insert(normalized).inserted else { return nil }
-      return (normalized as NSString).abbreviatingWithTildeInPath
-    }
-  }
-
-  static func isPaneZoomed<Surface: NSView & Identifiable>(
-    focusedSurfaceID: UUID?,
-    in tree: SplitTree<Surface>?
-  ) -> Bool where Surface.ID == UUID {
-    guard
-      let focusedSurfaceID,
-      let zoomedSurfaceID = tree?.zoomed?.leftmostLeaf().id
-    else {
-      return false
-    }
-    return focusedSurfaceID == zoomedSurfaceID
-  }
-
-  static func resolvedPaneDisplayTitle(
-    titleOverride: String?,
-    title: String?,
-    pwd: String?,
-    defaultValue: String
-  ) -> String {
-    if let titleOverride {
-      return titleOverride
-    }
-    if let title, !title.isEmpty {
-      return title
-    }
-    if let pwd = trimmedNonEmpty(pwd) {
-      return pwd
-    }
-    return defaultValue
-  }
-
-  static func resolvedTabDisplayTitle(
-    titleOverride: String?,
-    title: String?,
-    pwd: String?,
-    defaultValue: String
-  ) -> String {
-    if let titleOverride {
-      return titleOverride
-    }
-    if let title = trimmedNonEmpty(title) {
-      var resolved = strippedLeadingWorkingDirectory(from: title, pwd: pwd) ?? title
-      while let stripped = strippedDuplicatedTrailingCommandSuffix(from: resolved),
-        stripped != resolved
-      {
-        resolved = stripped
-      }
-      return resolved
-    }
-    if let pwd = trimmedNonEmpty(pwd) {
-      return pwd
-    }
-    return defaultValue
-  }
-
-  static func paneFallbackTitle<Surface: NSView & Identifiable>(
-    for surfaceID: UUID?,
-    in tree: SplitTree<Surface>?
-  ) -> String where Surface.ID == UUID {
-    let leaves = tree?.leaves() ?? []
-    guard !leaves.isEmpty else { return "Pane" }
-    if let surfaceID, let index = leaves.firstIndex(where: { $0.id == surfaceID }) {
-      return "Pane \(index + 1)"
-    }
-    return "Pane 1"
-  }
-
-  private static func isRunning(
-    progressState: ghostty_action_progress_report_state_e?
-  ) -> Bool {
-    switch progressState {
-    case .some(GHOSTTY_PROGRESS_STATE_SET),
-      .some(GHOSTTY_PROGRESS_STATE_INDETERMINATE),
-      .some(GHOSTTY_PROGRESS_STATE_PAUSE),
-      .some(GHOSTTY_PROGRESS_STATE_ERROR):
-      return true
-    default:
-      return false
-    }
-  }
-
-  private static func progressStateDescription(
-    _ progressState: ghostty_action_progress_report_state_e?
-  ) -> String? {
-    switch progressState {
-    case .some(GHOSTTY_PROGRESS_STATE_SET):
-      return "set"
-    case .some(GHOSTTY_PROGRESS_STATE_INDETERMINATE):
-      return "indeterminate"
-    case .some(GHOSTTY_PROGRESS_STATE_PAUSE):
-      return "pause"
-    case .some(GHOSTTY_PROGRESS_STATE_ERROR):
-      return "error"
-    case .some(GHOSTTY_PROGRESS_STATE_REMOVE):
-      return "remove"
-    default:
-      return nil
-    }
-  }
-
-  static func sidebarTerminalProgress(
-    state: GhosttySurfaceState?
-  ) -> TerminalSidebarTerminalProgress? {
-    guard let state else { return nil }
-
-    switch state.progressState {
-    case .some(GHOSTTY_PROGRESS_STATE_SET):
-      return .init(
-        fraction: state.progressValue.map { Double(Swift.max(0, Swift.min($0, 100))) / 100 },
-        tone: .active
-      )
-    case .some(GHOSTTY_PROGRESS_STATE_INDETERMINATE):
-      return .init(fraction: nil, tone: .active)
-    case .some(GHOSTTY_PROGRESS_STATE_PAUSE):
-      return .init(
-        fraction: state.progressValue.map { Double(Swift.max(0, Swift.min($0, 100))) / 100 } ?? 1,
-        tone: .paused
-      )
-    case .some(GHOSTTY_PROGRESS_STATE_ERROR):
-      return .init(
-        fraction: state.progressValue.map { Double(Swift.max(0, Swift.min($0, 100))) / 100 },
-        tone: .error
-      )
-    default:
-      return nil
-    }
-  }
-
-  private func updateRecentStructuredNotificationIfNeeded(
-    body: String,
-    createdAt: Date,
-    origin: NotificationOrigin,
-    surfaceID: UUID,
-    title: String
-  ) {
-    guard case .structuredAgent(let semantic) = origin else { return }
-    guard
-      let text = Self.normalizedNotificationText(
-        Self.notificationText(body: body, title: title)
-      )
-    else {
-      recentStructuredNotificationsBySurfaceID.removeValue(forKey: surfaceID)
-      return
-    }
-    recentStructuredNotificationsBySurfaceID[surfaceID] = .init(
-      recordedAt: createdAt,
-      semantic: semantic,
-      text: text
-    )
-  }
-
-  private func coalesceStructuredNotificationIfNeeded(
-    body: String,
-    origin: NotificationOrigin,
-    surfaceID: UUID,
-    title: String
-  ) {
-    guard case .structuredAgent(let semantic) = origin else { return }
-    guard
-      let structuredText = Self.normalizedNotificationText(
-        Self.notificationText(body: body, title: title)
-      ),
-      var notifications = paneNotifications[surfaceID]
-    else {
-      return
-    }
-    let now = Date()
-    guard
-      let index = notifications.indices.reversed().first(where: { index in
-        let notification = notifications[index]
-        guard
-          notification.origin == .terminalDesktop,
-          now.timeIntervalSince(notification.createdAt) <= Self.notificationCoalescingWindow,
-          let terminalText = Self.normalizedNotificationText(Self.notificationText(notification))
-        else {
-          return false
-        }
-        return Self.shouldCoalesceTerminalNotification(
-          terminalText: terminalText,
-          structuredText: structuredText,
-          semantic: semantic
-        )
-      })
-    else {
-      return
-    }
-    notifications.remove(at: index)
-    if notifications.isEmpty {
-      paneNotifications.removeValue(forKey: surfaceID)
-    } else {
-      paneNotifications[surfaceID] = notifications
-    }
-  }
-
-  private func shouldSuppressDesktopNotification(
-    body: String,
-    surfaceID: UUID,
-    title: String
-  ) -> Bool {
-    guard
-      let terminalText = Self.normalizedNotificationText(
-        Self.notificationText(body: body, title: title)
-      ),
-      let recentStructuredNotification = recentStructuredNotification(for: surfaceID)
-    else {
-      return false
-    }
-    return Self.shouldCoalesceTerminalNotification(
-      terminalText: terminalText,
-      structuredText: recentStructuredNotification.text,
-      semantic: recentStructuredNotification.semantic
-    )
-  }
-
-  private func recentStructuredNotification(for surfaceID: UUID) -> RecentStructuredNotification? {
-    guard let notification = recentStructuredNotificationsBySurfaceID[surfaceID] else {
-      return nil
-    }
-    guard Date().timeIntervalSince(notification.recordedAt) <= Self.notificationCoalescingWindow
-    else {
-      recentStructuredNotificationsBySurfaceID.removeValue(forKey: surfaceID)
-      return nil
-    }
-    return notification
-  }
-
-  static func latestNotification(in notifications: [PaneNotification]) -> PaneNotification? {
-    notifications.max { lhs, rhs in
-      lhs.createdAt < rhs.createdAt
-    }
-  }
-
-  static func unreadNotificationRecordCount(in notifications: [PaneNotification]) -> Int {
-    notifications.filter { $0.attentionState == .unread }.count
-  }
-
-  static func surfaceAttentionState(
-    in notifications: [PaneNotification]
-  ) -> SupatermNotificationAttentionState? {
-    if notifications.contains(where: { $0.attentionState == .unread }) {
-      return .unread
-    }
-    return nil
-  }
-
-  static func notificationsAfterDirectInteraction(
-    _ notifications: [PaneNotification],
-    activity: SurfaceActivity
-  ) -> [PaneNotification] {
-    guard activity.isFocused else { return notifications }
-    return notifications.map { notification in
-      guard notification.attentionState != nil else { return notification }
-      var updatedNotification = notification
-      updatedNotification.attentionState = nil
-      return updatedNotification
-    }
-  }
-
-  static func notificationText(_ notification: PaneNotification?) -> String? {
-    guard let notification else { return nil }
-    return notificationText(body: notification.body, title: notification.title)
-  }
-
-  static func sidebarNotificationPresentation(
-    _ notification: PaneNotification?
-  ) -> SidebarNotificationPresentation? {
-    guard let markdown = notificationText(notification) else { return nil }
-    return .init(
-      markdown: markdown,
-      previewMarkdown: sidebarNotificationPreviewMarkdown(markdown)
-    )
-  }
-
-  static func notificationText(body: String, title: String) -> String? {
-    let body = body.trimmingCharacters(in: .whitespacesAndNewlines)
-    if !body.isEmpty {
-      return body
-    }
-    let title = title.trimmingCharacters(in: .whitespacesAndNewlines)
-    return title.isEmpty ? nil : title
-  }
-
-  static func sidebarNotificationPreviewMarkdown(
-    _ markdown: String
-  ) -> String {
-    var preview = markdown.trimmingCharacters(in: .whitespacesAndNewlines)
-    let replacements = [
-      (#"(?m)^\s*\[[^\]]+\]:\s+\S.*$"#, ""),
-      (#"(?m)^\s*(```|~~~).*$"#, ""),
-      (#"(?m)^\s{0,3}#{1,6}\s*"#, ""),
-      (#"(?m)^\s{0,3}>\s?"#, ""),
-      (#"(?m)^\s*[-+*]\s+"#, ""),
-      (#"(?m)^\s*\d+[.)]\s+"#, ""),
-      (#"(?m)^\s*\[[ xX]\]\s+"#, ""),
-      (#"!\[([^\]]*)\]\([^)]+\)"#, "$1"),
-      (#"\[([^\]]+)\]\([^)]+\)"#, "$1"),
-      (#"\[([^\]]+)\]\[[^\]]*\]"#, "$1"),
-      (#"(?i)<(?:https?://|mailto:)[^>]+>"#, ""),
-      (#"(?i)\b(?:https?://|mailto:)\S+\b"#, ""),
-      (#"(?m)^\s*\|?(?:\s*:?-{3,}:?\s*\|)+\s*:?-{3,}:?\s*\|?\s*$"#, ""),
-    ]
-
-    for (pattern, template) in replacements {
-      preview = replacingMatches(in: preview, pattern: pattern, with: template)
-    }
-
-    preview =
-      preview
-      .split(separator: "\n", omittingEmptySubsequences: true)
-      .map { normalizeSidebarNotificationPreviewLine(String($0)) }
-      .filter { !$0.isEmpty }
-      .joined(separator: " ")
-
-    preview = replacingMatches(in: preview, pattern: #"\s+"#, with: " ", options: [])
-    return preview.trimmingCharacters(in: .whitespacesAndNewlines)
-  }
-
-  private static func normalizeSidebarNotificationPreviewLine(
-    _ line: String
-  ) -> String {
-    let trimmed = line.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return "" }
-
-    let pipeCount = trimmed.reduce(into: 0) { count, character in
-      if character == "|" {
-        count += 1
-      }
-    }
-    guard trimmed.hasPrefix("|") || trimmed.hasSuffix("|") || pipeCount >= 2 else {
-      return trimmed
-    }
-
-    let cells =
-      trimmed
-      .split(separator: "|", omittingEmptySubsequences: false)
-      .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
-      .filter { !$0.isEmpty }
-    return cells.joined(separator: " · ")
-  }
-
-  private static func replacingMatches(
-    in string: String,
-    pattern: String,
-    with template: String,
-    options: NSRegularExpression.Options = [.anchorsMatchLines]
-  ) -> String {
-    guard let expression = try? NSRegularExpression(pattern: pattern, options: options) else {
-      return string
-    }
-    let range = NSRange(string.startIndex..<string.endIndex, in: string)
-    return expression.stringByReplacingMatches(
-      in: string, options: [], range: range, withTemplate: template)
-  }
-
-  private static let genericCompletionNotificationTexts: Set<String> = [
-    "agent turn complete",
-    "task complete",
-    "turn complete",
-  ]
-
-  private static let notificationCoalescingWindow: TimeInterval = 2
-
-  private static func shouldCoalesceTerminalNotification(
-    terminalText: String,
-    structuredText: String,
-    semantic: NotificationSemantic
-  ) -> Bool {
-    if terminalText == structuredText {
-      return true
-    }
-    if terminalText.count < structuredText.count,
-      structuredText.hasPrefix(terminalText)
-    {
-      return true
-    }
-    return semantic == .completion
-      && genericCompletionNotificationTexts.contains(terminalText)
-  }
-
-  private static func normalizedNotificationText(_ value: String?) -> String? {
-    guard let value else { return nil }
-    let collapsed =
-      value
-      .components(separatedBy: .whitespacesAndNewlines)
-      .filter { !$0.isEmpty }
-      .joined(separator: " ")
-      .lowercased()
-      .trimmingCharacters(in: .punctuationCharacters)
-    return collapsed.isEmpty ? nil : collapsed
-  }
-
-  private static func trimmedNonEmpty(_ value: String?) -> String? {
-    let trimmed = value?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    return trimmed.isEmpty ? nil : trimmed
-  }
-
-  private static func strippedLeadingWorkingDirectory(
-    from title: String,
-    pwd: String?
-  ) -> String? {
-    guard let pwd = trimmedNonEmpty(pwd) else { return nil }
-    let normalizedPath = GhosttySurfaceView.normalizedWorkingDirectoryPath(pwd)
-    let abbreviatedPath = (normalizedPath as NSString).abbreviatingWithTildeInPath
-    let prefixes = Set([normalizedPath, abbreviatedPath]).sorted { $0.count > $1.count }
-
-    for prefix in prefixes where title.hasPrefix(prefix) {
-      let suffix = String(title.dropFirst(prefix.count))
-      guard let strippedSuffix = strippedTitleSeparator(from: suffix) else { continue }
-      return strippedSuffix
-    }
-
-    return nil
-  }
-
-  private static func strippedTitleSeparator(
-    from value: String
-  ) -> String? {
-    let separatorCharacters =
-      CharacterSet.whitespacesAndNewlines
-      .union(.punctuationCharacters)
-      .union(CharacterSet(charactersIn: "·•›»—–"))
-    guard let firstScalar = value.unicodeScalars.first,
-      separatorCharacters.contains(firstScalar)
-    else {
-      return nil
-    }
-
-    let stripped = String(
-      value.unicodeScalars.drop(while: { separatorCharacters.contains($0) })
-    ).trimmingCharacters(in: .whitespacesAndNewlines)
-
-    return stripped.isEmpty ? nil : stripped
-  }
-
-  private static func strippedDuplicatedTrailingCommandSuffix(
-    from title: String
-  ) -> String? {
-    guard let separatorRange = title.range(of: " - ", options: .backwards) else { return nil }
-    let prefix = String(title[..<separatorRange.lowerBound]).trimmingCharacters(
-      in: .whitespacesAndNewlines)
-    let suffix = String(title[separatorRange.upperBound...]).trimmingCharacters(
-      in: .whitespacesAndNewlines)
-    guard
-      let prefixCommand = leadingCommandToken(in: prefix),
-      let suffixCommand = trimmedNonEmpty(suffix),
-      prefixCommand == suffixCommand
-    else {
-      return nil
-    }
-    return prefix
-  }
-
-  private static func leadingCommandToken(
-    in value: String
-  ) -> String? {
-    guard let trimmed = trimmedNonEmpty(value) else { return nil }
-    return trimmed.split(whereSeparator: \.isWhitespace).first.map(String.init)
-  }
-}
-
-extension GhosttySurfaceView {
-  fileprivate var needsCloseConfirmation: Bool {
-    guard let surface else { return false }
-    return ghostty_surface_needs_confirm_quit(surface)
-  }
-
-  fileprivate func resolvedDisplayTitle(defaultValue: String) -> String {
-    TerminalHostState.resolvedPaneDisplayTitle(
-      titleOverride: bridge.state.titleOverride,
-      title: bridge.state.title,
-      pwd: bridge.state.pwd,
-      defaultValue: defaultValue
-    )
-  }
 }
